@@ -11,80 +11,80 @@
   },
   "homepage": "https://github.com/Submitty/Submitty#readme",
   "dependencies": {
-    "@codemirror/autocomplete": "6.13.0",
-    "@codemirror/commands": "6.3.2",
+    "@codemirror/autocomplete": "6.18.3",
+    "@codemirror/commands": "6.7.1",
     "@codemirror/lang-cpp": "6.0.2",
     "@codemirror/lang-java": "6.0.1",
     "@codemirror/lang-json": "6.0.1",
     "@codemirror/lang-python": "6.1.6",
-    "@codemirror/language": "6.10.1",
+    "@codemirror/language": "6.10.8",
     "@codemirror/language-data": "6.5.1",
-    "@codemirror/legacy-modes": "6.4.0",
-    "@codemirror/lint": "6.8.1",
-    "@codemirror/search": "6.5.6",
+    "@codemirror/legacy-modes": "6.4.2",
+    "@codemirror/lint": "6.8.4",
+    "@codemirror/search": "6.5.8",
     "@codemirror/state": "6.4.1",
     "@codemirror/theme-one-dark": "6.1.2",
-    "@codemirror/view": "6.23.0",
-    "@fortawesome/fontawesome-free": "6.4.2",
+    "@codemirror/view": "6.35.0",
+    "@fortawesome/fontawesome-free": "6.7.2",
     "@highlightjs/cdn-assets": "11.9.0",
     "@submitty/pdf-annotate.js": "2.5.1",
     "bootstrap": "4.6.0",
     "chosen-js": "1.8.7",
-    "codemirror": "5.65.16",
+    "codemirror": "5.65.18",
     "codemirror-spell-checker": "1.1.2",
     "flatpickr": "4.6.13",
     "jquery": "3.7.1",
-    "jquery-ui-dist": "1.13.2",
+    "jquery-ui-dist": "1.13.3",
     "jquery.are-you-sure": "1.9.0",
     "js-cookie": "3.0.5",
-    "jspdf": "^2.5.1",
-    "luxon": "3.4.4",
+    "jspdf": "^2.5.2",
+    "luxon": "3.5.0",
     "mermaid": "10.9.1",
     "pdfjs-dist": "2.10.377",
-    "plotly.js-dist": "2.34.0",
+    "plotly.js-dist": "2.35.2",
     "select2": "4.1.0-rc.0",
     "select2-bootstrap-5-theme": "1.3.0",
     "shortcut-buttons-flatpickr": "0.4.0",
-    "vue": "^3.4.21",
+    "vue": "^3.5.13",
     "twig": "1.17.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.25.2",
-    "@babel/preset-env": "^7.24.3",
-    "@babel/preset-typescript": "^7.22.11",
+    "@babel/core": "^7.26.0",
+    "@babel/preset-env": "^7.26.0",
+    "@babel/preset-typescript": "^7.26.0",
     "@cypress/skip-test": "^2.6.1",
     "@tsconfig/node20": "^20.1.4",
-    "@stylistic/eslint-plugin": "^2.1.0",
-    "@types/jest": "^29.5.12",
-    "@types/jquery": "^3.5.30",
+    "@stylistic/eslint-plugin": "^2.12.1",
+    "@types/jest": "^29.5.14",
+    "@types/jquery": "^3.5.32",
     "@types/jquery.are-you-sure": "^0.0.34",
-    "@types/jqueryui": "^1.12.22",
-    "@types/node": "^20.14.2",
+    "@types/jqueryui": "^1.12.23",
+    "@types/node": "^22.10.3",
     "@types/twig": "^1.12.16",
     "babel-jest": "^29.7.0",
-    "cypress": "^13.13.2",
+    "cypress": "^13.16.0",
     "cypress-browser-permissions": "^1.1.0",
     "cypress-file-upload": "^5.0.8",
-    "esbuild": "^0.21.1",
+    "esbuild": "^0.24.0",
     "eslint": "^8.57.0",
-    "eslint-plugin-cypress": "^3.3.0",
-    "eslint-plugin-jest": "^28.6.0",
-    "globals": "^15.4.0",
+    "eslint-plugin-cypress": "^3.6.0",
+    "eslint-plugin-jest": "^28.10.0",
+    "globals": "^15.13.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "node-fetch": "^2.6.12",
+    "node-fetch": "^2.7.0",
     "regenerator-runtime": "^0.14.1",
     "stylelint": "^15.11.0",
-    "stylelint-config-standard": "^33.0.0",
-    "typescript": "^5.4.5",
-    "typescript-eslint": "^7.13.0",
-    "vnu-jar": "^23.4.11",
-    "@vitejs/plugin-vue": "^5.0.4",
+    "stylelint-config-standard": "^34.0.0",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.19.0",
+    "vnu-jar": "^24.10.17",
+    "@vitejs/plugin-vue": "^5.2.1",
     "@vue/eslint-config-typescript": "^13.0.0",
-    "@vue/tsconfig": "^0.5.1",
-    "eslint-plugin-vue": "^9.23.0",
-    "vite": "^5.2.8",
-    "vue-tsc": "^2.0.11"
+    "@vue/tsconfig": "^0.7.0",
+    "eslint-plugin-vue": "^9.32.0",
+    "vite": "^5.4.10",
+    "vue-tsc": "^2.1.10"
   },
   "scripts": {
     "css-stylelint": "stylelint \"**/*.css\"",
