@@ -1,5 +1,6 @@
 #!/usr/bin/env python3 -u
 
+import sys
 import os
 from os import path
 import json
@@ -23,6 +24,7 @@ DAEMON_UID = OPEN_JSON['daemon_uid']
 with open(os.path.join(SUBMITTY_CONFIG_PATH)) as open_file:
     SUBMITTY_CONFIG = json.load(open_file)
 SUBMITTY_INSTALL_DIR = SUBMITTY_CONFIG['submitty_install_dir']
+SUBMITTY_REPOSITORY_DIR = SUBMITTY_CONFIG['submitty_repository']
 
 SYSTEMCTL_WRAPPER_SCRIPT = os.path.join(SUBMITTY_INSTALL_DIR, 'sbin', 'shipper_utils','systemctl_wrapper.py')
 
@@ -52,14 +54,48 @@ def update_docker_images(user, host, worker, autograding_workers, autograding_co
     if host == "localhost":
         get_sysinfo.print_distribution()
         client = docker.from_env()
+        try:
+            # List all images
+            image_set = {
+                image_name
+                for image in client.images.list()
+                for image_name in image.attrs["RepoTags"]
+            }
+
+            images_to_remove = set.difference(image_set, images_to_update)
+
+            # Prevent removal of system docker containers
+            with open(os.path.join(SUBMITTY_REPOSITORY_DIR, ".setup", "data", "system_docker_containers.json")) as json_file:
+                system_docker_containers = json.load(json_file)
+
+            images_to_remove = set.difference(images_to_remove, set(system_docker_containers))
+
+            # Remove images
+            for imageRemoved in images_to_remove:
+                try:
+                    image_id = client.images.get(imageRemoved).id
+                    print("Removed image " + imageRemoved)
+                except docker.errors.ImageNotFound as e:
+                    print(f"ERROR: Couldn't find image {imageRemoved}", file=sys.stderr)
+                    continue
+                try:
+                    client.images.remove(image_id, True)
+                except Exception as e:
+                    print(f"ERROR: An error occurred while removing image by ID {image_id}: {e}", file=sys.stderr)
+                    traceback.print_exc(file=sys.stderr)
+
+        except Exception as e:
+            print(f"ERROR: An error occurred: {e}", file=sys.stderr)
+            traceback.print_exc(file=sys.stderr)
+
         for image in images_to_update:
             print(f"locally pulling the image '{image}'")
             try:
                 repo, tag = image.split(':')
                 client.images.pull(repository=repo, tag=tag)
             except Exception as e:
-              print(f"ERROR: Could not pull {image}")
-              traceback.print_exc()
+              print(f"ERROR: Could not pull {image}: {e}", file=sys.stderr)
+              traceback.print_exc(file=sys.stderr)
 
               # check for machine
               if platform.machine() == "aarch64":
