@@ -18,7 +18,7 @@ class AutogradingStatusController extends AbstractController {
      */
     #[Route("/autograding_status", methods: ["GET"])]
     public function getGradingDonePage(): ResponseInterface {
-        if (empty($this->core->getQueries()->getInstructorLevelAccessCourse($this->core->getUser()->getId()))) {
+        if (empty($this->core->getQueries()->getInstructorLevelAccessCourse($this->core->getUser()->getId())) && !$this->core->getUser()->isSuperUser()) {
             return new MultiResponse(
                 JsonResponse::getFailResponse("You don't have access to this endpoint."),
                 new WebResponse(ErrorView::class, "errorPage", "You don't have access to this page.")
@@ -38,7 +38,7 @@ class AutogradingStatusController extends AbstractController {
      */
     #[Route("/autograding_status/get_update", methods: ["GET"])]
     public function getProgress(): JsonResponse {
-        if (empty($this->core->getQueries()->getInstructorLevelAccessCourse($this->core->getUser()->getId()))) {
+        if (empty($this->core->getQueries()->getInstructorLevelAccessCourse($this->core->getUser()->getId())) && !$this->core->getUser()->isSuperUser()) {
             return JsonResponse::getFailResponse("You do not have access to this endpoint.");
         }
         $info = $this->getAutogradingInfo();
@@ -51,7 +51,7 @@ class AutogradingStatusController extends AbstractController {
      */
     #[Route("/autograding_status/get_stack", methods: ["GET"])]
     public function getErrorStackTrace(): JsonResponse {
-        if (empty($this->core->getQueries()->getInstructorLevelAccessCourse($this->core->getUser()->getId()))) {
+        if (empty($this->core->getQueries()->getInstructorLevelAccessCourse($this->core->getUser()->getId())) && !$this->core->getUser()->isSuperUser()) {
             return JsonResponse::getFailResponse("You do not have access to this endpoint.");
         }
         $stack_trace_path = FileUtils::joinPaths($this->core->getConfig()->getSubmittyPath(), "logs", "autograding_stack_traces");
