@@ -77,6 +77,7 @@ class AdminGradeableController extends AbstractController {
             'using_subdirectory' => 'false',
             'vcs_subdirectory' => '',
             'syllabus_bucket' => 'Homework',
+            'autograding_config_path' => ''
         ];
 
         if (!isset($_POST['id']) || !isset($_POST['title']) || !isset($_POST['type'])) {
@@ -86,6 +87,7 @@ class AdminGradeableController extends AbstractController {
         $values['id'] = $_POST['id'];
         $values['title'] = $_POST['title'];
         $values['type'] = $_POST['type'];
+        $values['autograding_config_path'] = $_POST['autograding_config_path'] ?? FileUtils::joinPaths($this->core->getConfig()->getSubmittyInstallPath(), 'more_autograding_examples/upload_only/config');
         if ($_POST['type'] === 'Electronic File') {
             if (array_key_exists('vcs', $_POST)) {
                 if (!array_key_exists('repository_type', $_POST['vcs'])) {
@@ -1386,7 +1388,15 @@ class AdminGradeableController extends AbstractController {
             if (isset($details[$date_property])) {
                 $dates[$date_property] = $details[$date_property];
 
-                if ($dates[$date_property] > DateUtils::MAX_TIME) {
+                if ($date_property === 'late_days') {
+                    if (!is_numeric($dates[$date_property])) {
+                        $errors[$date_property] = 'Late days must be a number';
+                    }
+                    elseif (intval($dates[$date_property]) < 0) {
+                        $errors[$date_property] = 'Late days must be a positive number';
+                    }
+                }
+                elseif ($dates[$date_property] > DateUtils::MAX_TIME) {
                     $errors[$date_property] = Gradeable::date_display_names[$date_property] . ' Date is higher than the max allowed date! (' . DateUtils::MAX_TIME . ')';
                 }
 
