@@ -36,14 +36,19 @@ class ErrorView extends AbstractView {
      * @param string $readd_url URL to the rejoin course function.
      * @return string The Twig HTML for this page.
      */
-    public function noAccessCourse(bool $can_rejoin_course, string $readd_url): string {
+    public function noAccessCourse(bool $can_rejoin_course, string $readd_url, int $self_registration_type): string {
         return $this->core->getOutput()->renderTwigTemplate("error/NoAccessCourse.twig", [
             "course_name" => $this->core->getDisplayedCourseName(),
             "semester" => $this->core->getFullSemester(),
+            "course_id" => $this->core->getConfig()->getCourse(),
             "main_url" => $this->core->getConfig()->getBaseUrl(),
+            "course_home_url" => $this->core->getConfig()->getCourseHomeUrl(),
             "ability_to_readd" => $can_rejoin_course,
             "readd_url" => $readd_url,
-            "csrf_token" => $this->core->getCsrfToken()
+            "register_url" => $this->core->buildCourseUrl(['register']),
+            "user" => $this->core->getUser(),
+            "csrf_token" => $this->core->getCsrfToken(),
+            "self_registration_type" => $self_registration_type
         ]);
     }
 
