@@ -24,14 +24,16 @@ class PollView extends AbstractView {
      *
      * @param Poll[] $todays_polls
      * @param Poll[] $older_polls
+     * @param Poll[] $tomorrow_polls
      * @param Poll[] $future_polls
      */
-    public function showPollsInstructor(array $todays_polls, array $older_polls, array $future_polls, array $response_counts, array $dropdown_states) {
+    public function showPollsInstructor(array $todays_polls, array $older_polls, array $tomorrow_polls, array $future_polls, array $response_counts, array $dropdown_states) {
         return $this->core->getOutput()->renderTwigTemplate("polls/AllPollsPageInstructor.twig", [
             'csrf_token' => $this->core->getCsrfToken(),
             'base_url' => $this->core->buildCourseUrl() . '/polls',
             'todays_polls' => $todays_polls,
             'older_polls' => $older_polls,
+            'tomorrow_polls' => $tomorrow_polls,
             'future_polls' => $future_polls,
             'dropdown_states' => $dropdown_states,
             'response_counts' => $response_counts,
