@@ -7,7 +7,8 @@ use app\views\AbstractView;
 use app\libraries\FileUtils;
 use app\libraries\ForumUtils;
 use app\models\User;
-use app\controllers\forum\ForumController;
+use app\entities\forum\Thread;
+use app\entities\forum\Post;
 
 class ForumThreadView extends AbstractView {
     private function getSavedForumCategories($current_course, $categories) {
@@ -157,19 +158,57 @@ class ForumThreadView extends AbstractView {
      * for a specific thread, in addition to head of the threads
      * that have been created after applying filter and to be
      * displayed in the left panel.
+     * @param Thread[] $threads
+     * @param Thread[] $merge_thread_options
+     * @return mixed[]|string
      */
-
-    public function showForumThreads($user, $posts, $unviewed_posts, $threadsHead, $show_deleted, $show_merged_thread, $display_option, $max_thread, $initialPageNumber, $thread_resolve_state, $post_content_limit, $ajax = false, $thread_announced = true) {
-        $threadExists = $this->core->getQueries()->threadExists();
-        $filteredThreadExists = (count($threadsHead) > 0);
-        $currentThread = -1;
-        $currentCategoriesIds = [];
-        $show_deleted_thread_title = null;
+    public function showForumThreads(string $user, Thread $thread, array $threads, array $merge_thread_options, bool $show_deleted, bool $show_merged_thread, string $display_option, int $initialPageNumber, bool $ajax = false): array|string {
         $currentCourse = $this->core->getConfig()->getCourse();
-        $threadFiltering = $threadExists && !$filteredThreadExists && !(empty($_COOKIE[$currentCourse . '_forum_categories']) && empty($_COOKIE['forum_thread_status']) && !empty($_COOKIE['unread_select_value']) && $_COOKIE['unread_select_value'] === 'false');
+        $categories = $this->core->getQueries()->getCategories();
+
+        $cookieSelectedCategories = $this->getSavedForumCategories($currentCourse, $categories);
+        $cookieSelectedThreadStatus = $this->getSavedThreadStatuses();
+        $cookieSelectedUnread = $this->getUnreadThreadStatus();
+
+        $filterFormData = [
+            "categories" => $categories,
+            "current_thread" => $thread->getId(),
+            "current_category_ids" => $thread->getCategories()->map(function ($x) {
+                return $x->getId();
+            }),
+            "current_course" => $currentCourse,
+            "cookie_selected_categories" => $cookieSelectedCategories,
+            "cookie_selected_thread_status" => $cookieSelectedThreadStatus,
+            "cookie_selected_unread_value" => $cookieSelectedUnread,
+            "display_option" => $display_option,
+            "thread_exists" => true,
+        ];
+
+        $next_page = $initialPageNumber + 1;
+        $prev_page = $initialPageNumber - 1;
+        $arrowup_visibility = ($initialPageNumber === 0) ? "display:none;" : "";
+        $displayThreadContent = $this->displayThreadList($threads, false);
+
+        $generatePostContent = $this->generatePostList($thread, true, $display_option, $merge_thread_options, false);
+
+        $this->core->getQueries()->visitThread($user, $thread->getId());
+
+        $return = "";
+
+        $markdown_enabled = 0;
+        if (isset($_COOKIE['markdown_enabled'])) {
+            $markdown_enabled = $_COOKIE['markdown_enabled'];
+        }
+
+        $button_params = $this->getAllForumButtons(true, $thread->getId(), $display_option, $show_deleted, $show_merged_thread);
 
         if (!$ajax) {
-            $this->core->getOutput()->addBreadcrumb("Discussion Forum", $this->core->buildCourseUrl(['forum']), null, $use_as_heading = true);
+            $this->core->getOutput()->addBreadcrumb("Discussion Forum", $this->core->buildCourseUrl(['forum']), null, true);
+             // Add breadcrumb for the current thread
+             $max_length = 25;
+             $fullTitle = $thread->getTitle();
+             $title = strlen($fullTitle) > $max_length ? substr($fullTitle, 0, $max_length - 3) . "..." : $fullTitle;
+             $this->core->getOutput()->addBreadcrumb("(" . $thread->getId() . ") " . $title, $this->core->buildCourseUrl(['forum', 'threads', $thread->getId()]), null, true);
 
             //Body Style is necessary to make sure that the forum is still readable...
             $this->core->getOutput()->addVendorCss('codemirror/codemirror.css');
@@ -189,98 +228,24 @@ class ForumThreadView extends AbstractView {
             $this->core->getOutput()->addInternalJs('forum.js');
             $this->core->getOutput()->addVendorJs('jquery.are-you-sure/jquery.are-you-sure.js');
             $this->core->getOutput()->addVendorJs('bootstrap/js/bootstrap.bundle.min.js');
-        }
 
-        if ($filteredThreadExists || $threadFiltering) {
-            $currentThread = isset($_GET["thread_id"]) && is_numeric($_GET["thread_id"]) && (int) $_GET["thread_id"] < $max_thread && (int) $_GET["thread_id"] > 0 ? (int) $_GET["thread_id"] : $posts[0]["thread_id"];
-            $currentCategoriesIds = $this->core->getQueries()->getCategoriesIdForThread($currentThread);
-        }
-
-        $currentThreadArr = array_filter($threadsHead, function ($ar) use ($currentThread) {
-            return ($ar['id'] == $currentThread);
-        });
-
-        $categories = $this->core->getQueries()->getCategories();
-
-        $cookieSelectedCategories = $this->getSavedForumCategories($currentCourse, $categories);
-        $cookieSelectedThreadStatus = $this->getSavedThreadStatuses();
-        $cookieSelectedUnread = $this->getUnreadThreadStatus();
-
-        $filterFormData = [
-            "categories" => $categories,
-            "current_thread" => $currentThread,
-            "current_category_ids" => $currentCategoriesIds,
-            "current_course" => $currentCourse,
-            "cookie_selected_categories" => $cookieSelectedCategories,
-            "cookie_selected_thread_status" => $cookieSelectedThreadStatus,
-            "cookie_selected_unread_value" => $cookieSelectedUnread,
-            "display_option" => $display_option,
-            "thread_exists" => $threadExists
-        ];
-
-        $next_page = 0;
-        $prev_page = 0;
-        $arrowup_visibility = 0;
-        $displayThreadContent = "";
-        $generatePostContent = "";
-
-        if ($threadExists) {
-            $next_page = $initialPageNumber + 1;
-            $prev_page = ($initialPageNumber == 1) ? 0 : ($initialPageNumber - 1);
-            $arrowup_visibility = ($initialPageNumber == 1) ? "display:none;" : "";
-            $activeThreadAnnouncement = false;
-            $activeThreadTitle = "";
-            $activeThread = [];
-            $displayThreadContent = $this->displayThreadList($threadsHead, false, $activeThreadAnnouncement, $activeThreadTitle, $activeThread, $currentThread, $currentCategoriesIds, false);
-
-            if (count($activeThread) == 0) {
-                $activeThread = $this->core->getQueries()->getThread($currentThread);
-            }
-
-            $currentThreadArrValues = array_values($currentThreadArr);
-            $currentThreadFavorite = !empty($currentThreadArrValues) ? $currentThreadArrValues[0]['favorite'] : false;
-            $generatePostContent = $this->generatePostList($currentThread, $posts, $unviewed_posts, $currentCourse, true, $threadExists, $display_option, $categories, $cookieSelectedCategories, $cookieSelectedThreadStatus, $cookieSelectedUnread, $currentCategoriesIds, $currentThreadFavorite, false, $thread_announced);
-        }
-
-        if (!empty($activeThread['id'])) {
-            $this->core->getQueries()->visitThread($user, $activeThread['id']);
-        }
-
-        $return = "";
-
-        $markdown_enabled = 0;
-        if (isset($_COOKIE['markdown_enabled'])) {
-            $markdown_enabled = $_COOKIE['markdown_enabled'];
-        }
-
-        $button_params = $this->getAllForumButtons($threadExists, $currentThread, $display_option, $show_deleted, $show_merged_thread);
-
-        if (!$ajax) {
-            // Add breadcrumb for the current thread
-            $currentThreadArrValues = array_values($currentThreadArr);
-            if ($currentThreadArrValues) {
-                $max_length = 25;
-                $fullTitle = $currentThreadArrValues[0]["title"];
-                $title = strlen($fullTitle) > $max_length ? substr($fullTitle, 0, $max_length - 3) . "..." : $fullTitle;
-                $this->core->getOutput()->addBreadcrumb("(" . $currentThreadArrValues[0]["id"] . ") " . $title, $this->core->buildCourseUrl(['forum', 'threads', 9]), null, $use_as_heading = true);
-            }
 
             $return = $this->core->getOutput()->renderTwigTemplate("forum/ShowForumThreads.twig", [
                 "categories" => $categories,
                 "filterFormData" => $filterFormData,
                 "button_params" => $button_params,
-                "thread_exists" => $threadExists,
+                "thread_exists" => true,
                 "next_page" => $next_page,
                 "prev_page" => $prev_page,
                 "arrowup_visibility" => $arrowup_visibility,
                 "display_thread_content" => $displayThreadContent,
-                "display_thread_count" => empty($displayThreadContent) ? 0 : count($displayThreadContent["thread_content"]),
-                "currentThread" => $currentThread,
+                "display_thread_count" => count($displayThreadContent["thread_content"]),
+                "currentThread" => $thread,
                 "currentCourse" => $currentCourse,
                 "accessGrading" => $this->core->getUser()->accessGrading(),
                 "manage_categories_url" => $this->core->buildCourseUrl(['forum', 'categories']),
                 "generate_post_content" => $generatePostContent,
-                "thread_resolve_state" => $thread_resolve_state,
+                "thread_resolve_state" => $thread->getStatus(),
                 "show_unresolve" => false,
                 "display_option" => $display_option,
                 "render_markdown" => $markdown_enabled,
@@ -289,7 +254,7 @@ class ForumThreadView extends AbstractView {
                 "search_url" => $this->core->buildCourseUrl(['forum', 'search']),
                 "merge_url" => $this->core->buildCourseUrl(['forum', 'threads', 'merge']),
                 "split_url" => $this->core->buildCourseUrl(['forum', 'posts', 'split']),
-                "post_content_limit" => $post_content_limit
+                "post_content_limit" => ForumUtils::FORUM_CHAR_POST_LIMIT
             ]);
         }
         else {
@@ -307,7 +272,7 @@ class ForumThreadView extends AbstractView {
                 "thread_id" => $generatePostContent["thread_id"],
                 "first_post_id" => $generatePostContent["first_post_id"],
                 "form_action_link" => $generatePostContent["form_action_link"],
-                "thread_resolve_state" => $thread_resolve_state,
+                "thread_resolve_state" => $thread->getStatus(),
                 "show_unresolve" => false,
                 "merge_thread_content" => $generatePostContent["merge_thread_content"],
                 "csrf_token" => $generatePostContent["csrf_token"],
@@ -315,7 +280,7 @@ class ForumThreadView extends AbstractView {
                 "post_box_id" => $generatePostContent["post_box_id"],
                 "merge_url" => $this->core->buildCourseUrl(['forum', 'threads', 'merge']),
                 "split_url" => $this->core->buildCourseUrl(['forum', 'posts', 'split']),
-                "post_content_limit" => $post_content_limit,
+                "post_content_limit" => ForumUtils::FORUM_CHAR_POST_LIMIT,
                 "render_markdown" => $markdown_enabled
             ]);
 
@@ -431,158 +396,70 @@ class ForumThreadView extends AbstractView {
         }
         return $button_params;
     }
-
-
-    public function generatePostList($currentThread, $posts, $unviewed_posts, $currentCourse, $includeReply = false, $threadExists = false, $display_option = 'time', $categories = [], $cookieSelectedCategories = [], $cookieSelectedThreadStatus = [], $cookieSelectedUnread = [], $currentCategoriesIds = [], $isCurrentFavorite = false, $render = true, $thread_announced = false) {
-
-        $activeThread = $this->core->getQueries()->getThread($currentThread);
-
-        $activeThreadTitle = "({$activeThread['id']}) " . $activeThread['title'];
-        $activeThreadAnnouncement = $activeThread['pinned_expiration'] > date("Y-m-d H:i:s");
-        $expiring = $activeThread['pinned_expiration'] <= date("Y-m-d H:i:s", strtotime("+7 day"));
-
-        $thread_id = $activeThread['id'];
-        if ($thread_id === -1) {
-            $thread_id = array_values($posts)[0]["thread_id"];
+    /**
+     * @param Post[] $reply_hierarchy
+     * @return Post[]
+     */
+    private function BuildReplyHeirarchy(Post $post, array &$reply_hierarchy = [], int $reply_level = 1): array {
+        $reply_hierarchy[] = $post;
+        $post->setReplyLevel($reply_level);
+        foreach ($post->getChildren() as $child) {
+            $this->BuildReplyHeirarchy($child, $reply_hierarchy, $reply_level + 1);
         }
-        $first_post = $this->core->getQueries()->getFirstPostForThread($thread_id);
-        $first_post_id = $first_post["id"];
-        $first_post_author_id = $first_post['author_user_id'];
-        $first_post_anonymous = ($first_post['anonymous'] === true);
+        return $reply_hierarchy;
+    }
 
+    /**
+     * Renders or formats a list of posts.
+     * @param Thread $thread
+     * @param bool $includeReply
+     * @param string $display_option
+     * @param Thread[] $merge_thread_options
+     * @param bool $render
+     * @return mixed[]|string
+     */
+    public function generatePostList(Thread $thread, bool $includeReply, string $display_option, array $merge_thread_options, bool $render = true): array|string {
         $first = true;
-
         $post_data = [];
-
         $csrf_token = $this->core->getCsrfToken();
-
-        $totalAttachments = 0;
         $GLOBALS['totalAttachments'] = 0;
-
-        $author_user_groups_map = [];
-
-        $author_user_ids = array_map(function ($post) {
-            return $post["author_user_id"];
-        }, $posts);
-
-        $author_user_groups = $this->core->getQueries()->getAuthorUserGroups($author_user_ids);
-        $authors_display_info = $this->core->getQueries()->getDisplayUserInfoFromUserIds($author_user_ids);
-
-        foreach ($author_user_groups as $author) {
-            $author_user_groups_map[$author["user_id"]] = $author["user_group"];
+        $user = $this->core->getUser();
+        $first_post = $thread->getFirstPost();
+        if ($first_post === false) {
+            $first_post = new Post($thread);
         }
 
-        $post_ids = array_column($posts, "id");
-        $posts_with_history = $this->core->getQueries()->getPostsWithHistory($post_ids);
-        $merged_threads = $this->core->getQueries()->getMergedThreadIds($post_ids);
-        $post_attachments = $this->core->getQueries()->getForumAttachments($post_ids);
-
-        $current_user = $this->core->getUser()->getId();
-        $upDuckCounter_map = [];
-        $upDuckCounter_map = $this->core->getQueries()->getUpduckInfoForPosts($post_ids);
-        $userLiked = $this->core->getQueries()->getUserLikesForPosts($post_ids, $current_user);
-        $staffLiked = $this->core->getQueries()->getInstructorUpduck($post_ids);
+        $posts = [];
         if ($display_option == "tree") {
-            $order_reply_level_array = ForumController::getPostsOrderAndReplies($posts, $thread_id);
-            $order_array = $order_reply_level_array[0];
-            $reply_level_array = $order_reply_level_array[1];
-            $i = 0;
-            $first = true;
-
-            foreach ($order_array as $ordered_post) {
-                foreach ($posts as $post) {
-                    if ($post["id"] == $ordered_post) {
-                        if ($post["parent_id"] == $first_post_id) {
-                            $reply_level = 1;
-                        }
-                        else {
-                            $reply_level = $reply_level_array[$i];
-                        }
-
-                        $post["author_user_group"] = $author_user_groups_map[$post["author_user_id"]];
-
-                        $boolLiked = in_array($post["id"], $userLiked, true);
-
-                        $boolStaffLiked = in_array($post["id"], $staffLiked, true);
-
-                        $post_data[] = $this->createPost(
-                            $first_post_author_id,
-                            $first_post_anonymous,
-                            $activeThread,
-                            $post,
-                            $unviewed_posts,
-                            $first,
-                            $reply_level,
-                            $display_option,
-                            $upDuckCounter_map[$post["id"]],
-                            $boolLiked,
-                            $boolStaffLiked,
-                            $includeReply,
-                            $authors_display_info[$post['author_user_id']],
-                            $post_attachments[$post["id"]][0],
-                            in_array($post["id"], $posts_with_history, true),
-                            in_array($post["id"], $merged_threads, true),
-                            false,
-                            $thread_announced,
-                            $isCurrentFavorite,
-                        );
-
-                        break;
-                    }
-                }
-                if ($first) {
-                    $first = false;
-                }
-                $i++;
-            }
+            $posts = $this->BuildReplyHeirarchy($first_post);
         }
         else {
-            foreach ($posts as $post) {
-                $post["author_user_group"] = $author_user_groups_map[$post["author_user_id"]];
-                $boolLiked = in_array($post["id"], $userLiked, true);
-                $boolStaffLiked = in_array($post["id"], $staffLiked, true);
-                $post_data[] = $this->createPost(
-                    $first_post_author_id,
-                    $first_post_anonymous,
-                    $activeThread,
-                    $post,
-                    $unviewed_posts,
-                    $first,
-                    1,
-                    $display_option,
-                    $upDuckCounter_map[$post["id"]],
-                    $boolLiked,
-                    $boolStaffLiked,
-                    $includeReply,
-                    $authors_display_info[$post['author_user_id']],
-                    $post_attachments[$post["id"]][0],
-                    in_array($post["id"], $posts_with_history, true),
-                    in_array($post["id"], $merged_threads, true),
-                    false,
-                    $thread_announced,
-                    $isCurrentFavorite,
-                );
-
-                if ($first) {
-                    $first = false;
-                }
-            }
+            // posts were ordered at query-time by repository
+            $posts = $thread->getPosts()->toArray();
         }
-
-        $isThreadLocked = $this->core->getQueries()->isThreadLocked($thread_id);
-        $accessFullGrading = $this->core->getUser()->accessFullGrading();
-
-        $post_box_id = 1;
+        $post_box_id = 2;
+        foreach ($posts as $post) {
+            $post_data[] = $this->createPost(
+                $first_post,
+                $thread,
+                $post,
+                $first,
+                $display_option,
+                $includeReply,
+                $post_box_id,
+                false,
+            );
+            if ($first) {
+                $first = false;
+            }
+            $post_box_id++;
+        }
 
         $form_action_link = $this->core->buildCourseUrl(['forum', 'posts', 'new']);
 
-        if (($isThreadLocked != 1 || $accessFullGrading ) && $includeReply) {
-            $GLOBALS['post_box_id'] = $post_box_id = isset($GLOBALS['post_box_id']) ? $GLOBALS['post_box_id'] + 1 : 2;
-        }
-
         $merge_thread_content = [];
 
-        if ($this->core->getUser()->getGroup() <= 3) {
+        if ($this->core->getUser()->accessGrading()) {
             $this->core->getOutput()->addVendorCss(FileUtils::joinPaths('chosen-js', 'chosen.min.css'));
             $this->core->getOutput()->addVendorJs(FileUtils::joinPaths('chosen-js', 'chosen.jquery.min.js'));
             $this->core->getOutput()->addVendorCss(FileUtils::joinPaths('flatpickr', 'flatpickr.min.css'));
@@ -592,90 +469,48 @@ class ForumThreadView extends AbstractView {
             $this->core->getOutput()->addInternalJs('autosave-utils.js');
             $this->core->getOutput()->addInternalJs('forum.js');
             $this->core->getOutput()->addInternalCss('forum.css');
-            $current_thread_date = $first_post["timestamp"];
-            $merge_thread_list = $this->core->getQueries()->getThreadsBefore($current_thread_date, 1);
-            $merge_thread_first_posts = $this->core->getQueries()->getFirstPostForThreads(array_column($merge_thread_list, "id"));
-            // Get first post of each thread. To be used later
-            // to obtain the content of the post to be displayed
-            // in the modal.
-            foreach ($merge_thread_list as $key => $temp_thread) {
-                $merge_thread_list[$key]['first_post_id'] = $merge_thread_first_posts[$temp_thread['id']]['id'];
-            }
-
             $merge_thread_content = [
-                "current_thread_date" => $current_thread_date,
-                "current_thread" => $currentThread,
-                "possibleMerges" => $merge_thread_list
+                "current_thread_date" => $first_post->getTimestamp(),
+                "current_thread" => $thread->getId(),
+                "possibleMerges" => $merge_thread_options
             ];
         }
 
-        $return = "";
-
+        $generated_post_list = [
+            "userGroup" => $user->getGroup(),
+            "activeThread" => $thread,
+            "activeThreadAnnouncement" => $thread->isPinned(),
+            "expiring" => $thread->isPinnedExpiring(),
+            "isCurrentFavorite" => $thread->isFavorite($user->getId()),
+            "display_option" => $display_option,
+            "post_data" => $post_data,
+            "isThreadLocked" => $thread->isLocked(),
+            "accessFullGrading" => $user->accessFullGrading(),
+            "includeReply" => $includeReply,
+            "thread_id" => $thread->getId(),
+            "first_post_id" => $first_post->getId(),
+            "form_action_link" => $form_action_link,
+            "merge_thread_content" => $merge_thread_content,
+            "csrf_token" => $csrf_token,
+            "activeThreadTitle" => "({$thread->getId()}) " . $thread->getTitle(),
+            "post_box_id" => $post_box_id,
+            "total_attachments" => $GLOBALS['totalAttachments'],
+            "merge_url" => $this->core->buildCourseUrl(['forum', 'threads', 'merge']),
+            "split_url" => $this->core->buildCourseUrl(['forum', 'posts', 'split'])
+        ];
         if ($render) {
-            $return = $this->core->getOutput()->renderTwigTemplate("forum/GeneratePostList.twig", [
-                "userGroup" => $this->core->getUser()->getGroup(),
-                "activeThread" => $activeThread,
-                "activeThreadAnnouncement" => $activeThreadAnnouncement,
-                "expiring" => $expiring,
-                "isCurrentFavorite" => $isCurrentFavorite,
-                "display_option" => $display_option,
-                "post_data" => $post_data,
-                "isThreadLocked" => $isThreadLocked,
-                "accessFullGrading" => $accessFullGrading,
-                "includeReply" => $includeReply,
-                "thread_id" => $thread_id,
-                "first_post_id" => $first_post_id,
-                "form_action_link" => $form_action_link,
-                "merge_thread_content" => $merge_thread_content,
-                "csrf_token" => $csrf_token,
-                "activeThreadTitle" => $activeThreadTitle,
-                "post_box_id" => $post_box_id,
-                "total_attachments" => $GLOBALS['totalAttachments'],
-                "merge_url" => $this->core->buildCourseUrl(['forum', 'threads', 'merge']),
-                "split_url" => $this->core->buildCourseUrl(['forum', 'posts', 'split'])
-            ]);
+            $generated_post_list = $this->core->getOutput()->renderTwigTemplate("forum/GeneratePostList.twig", $generated_post_list);
         }
-        else {
-            $return = [
-                "userGroup" => $this->core->getUser()->getGroup(),
-                "activeThread" => $activeThread,
-                "activeThreadAnnouncement" => $activeThreadAnnouncement,
-                "expiring" => $expiring,
-                "isCurrentFavorite" => $isCurrentFavorite,
-                "display_option" => $display_option,
-                "post_data" => $post_data,
-                "isThreadLocked" => $isThreadLocked,
-                "accessFullGrading" => $accessFullGrading,
-                "includeReply" => $includeReply,
-                "thread_id" => $thread_id,
-                "first_post_id" => $first_post_id,
-                "form_action_link" => $form_action_link,
-                "merge_thread_content" => $merge_thread_content,
-                "csrf_token" => $csrf_token,
-                "activeThreadTitle" => $activeThreadTitle,
-                "post_box_id" => $post_box_id,
-                "total_attachments" => $GLOBALS['totalAttachments']
-            ];
-        }
-
-        return $return;
+        return $generated_post_list;
     }
 
-    public function showAlteredDisplayList($threads, $filtering, $thread_id, $categories_ids, $ajax = false) {
-        $tempArray = [];
-        $threadAnnouncement = false;
-        $activeThreadTitle = "";
-        $thread = "";
-        if ($ajax) {
-            for ($i = 0; $i < count($threads); $i++) {
-                if ($threads[$i]["id"] == $thread_id) {
-                    $thread = $threads[$i];
-                    break;
-                }
-            }
-            $threads = [$thread];
-        }
-        return $this->displayThreadList($threads, $filtering, $threadAnnouncement, $activeThreadTitle, $tempArray, $thread_id, $categories_ids, true);
+    /**
+     * Renders scroll content or new posts from websocket.
+     * @param Thread[] $threads
+     * @return string
+     */
+    public function showAlteredDisplayList(array $threads): string {
+        return $this->displayThreadList($threads, true, true);
     }
 
     public function contentMarkdownToPlain($str) {
@@ -685,23 +520,20 @@ class ForumThreadView extends AbstractView {
         return $str;
     }
 
-    public function showFullThreadsPage($threads, $category_ids, $show_deleted, $show_merged_threads, $page_number) {
-        $activeThreadAnnouncements = [];
-        $activeThreadTitle = "";
-        $activeThread = [];
+    /**
+     * Renders the main forum page
+     * @param Thread[] $threads
+     * @param bool $show_deleted
+     * @param bool $show_merged_threads
+     * @return string
+     */
+    public function showFullThreadsPage(array $threads, bool $show_deleted, bool $show_merged_threads, int $block_number): string {
         $GLOBALS['totalAttachments'] = 0;
-        $thread_content =  $this->displayThreadList($threads, false, $activeThreadAnnouncements, $activeThreadTitle, $activeThread, null, $category_ids, false, true);
+        $thread_content =  $this->displayThreadList($threads, false, true);
         $categories = $this->core->getQueries()->getCategories();
         $current_course = $this->core->getConfig()->getCourse();
-        $cookieSelectedCategories = $this->getSavedForumCategories($current_course, $categories);
-        $cookieSelectedThreadStatus = $this->getSavedThreadStatuses();
-        $cookieSelectedUnread = $this->getUnreadThreadStatus();
-        $next_page = 0;
-        $prev_page = 0;
         // getting the forum page buttons
-        $thread_id = -1;
-        $thread_exists = $this->core->getQueries()->threadExists();
-        $button_params = $this->getAllForumButtons($thread_exists, $thread_id, null, $show_deleted, $show_merged_threads);
+        $button_params = $this->getAllForumButtons(true, -1, null, $show_deleted, $show_merged_threads);
 
         // add css and js files
         $this->core->getOutput()->addInternalCss("forum.css");
@@ -716,20 +548,15 @@ class ForumThreadView extends AbstractView {
         $this->core->getOutput()->addVendorCss('flatpickr/flatpickr.min.css');
         $this->core->getOutput()->addVendorCss(FileUtils::joinPaths('flatpickr', 'plugins', 'shortcutButtons', 'themes', 'light.min.css'));
 
-        if ($thread_exists) {
-            $next_page = $page_number + 1;
-            $prev_page = $page_number - 1;
-        }
-
         $filterFormData = [
             "categories" => $categories,
-            "current_thread" => $thread_id,
+            "current_thread" => -1,
             "current_category_ids" => [],
             "current_course" => $current_course,
-            "cookie_selected_categories" => $cookieSelectedCategories,
-            "cookie_selected_thread_status" => $cookieSelectedThreadStatus,
-            "cookie_selected_unread_value" => $cookieSelectedUnread,
-            "thread_exists" => $thread_exists
+            "cookie_selected_categories" => $this->getSavedForumCategories($current_course, $categories),
+            "cookie_selected_thread_status" => $this->getSavedThreadStatuses(),
+            "cookie_selected_unread_value" => $this->getUnreadThreadStatus(),
+            "thread_exists" => true
         ];
 
 
@@ -737,8 +564,8 @@ class ForumThreadView extends AbstractView {
             "thread_content" => $thread_content["thread_content"],
             "button_params" => $button_params,
             "filterFormData" => $filterFormData,
-            "next_page" => $next_page,
-            "prev_page" => $prev_page,
+            "next_page" => count($threads) > 0 ? $block_number + 1 : 0,
+            "prev_page" => count($threads) > 0 ? $block_number - 1 : 0,
             "display_thread_count" => empty($thread_content) ? 0 : count($thread_content["thread_content"]),
             "csrf_token" => $this->core->getCsrfToken(),
             "search_url" => $this->core->buildCourseUrl(['forum', 'search']),
@@ -746,13 +573,13 @@ class ForumThreadView extends AbstractView {
             "edit_url" => $this->core->buildCourseUrl(['forum']),
             "current_user" => $this->core->getUser()->getId(),
             "user_group" => $this->core->getUser()->getGroup(),
-            "thread_exists" => $thread_exists,
+            "thread_exists" => true,
             "manage_categories_url" => $this->core->buildCourseUrl(['forum', 'categories'])
         ]);
     }
 
-    public function sizeTitle($titleDisplay, $title, $titleLength, $length = 40) {
-        $titleDisplay = substr($titleDisplay, 0, ($titleLength < $length) ? $titleLength : strrpos(substr($titleDisplay, 0, $length), " "));
+    public function sizeTitle(string $title, int $titleLength, $length = 40) {
+        $titleDisplay = substr($title, 0, ($titleLength < $length) ? $titleLength : strrpos(substr($title, 0, $length), " "));
 
         if ($titleLength > $length) {
             //Fix ... appearing
@@ -772,86 +599,53 @@ class ForumThreadView extends AbstractView {
         return $contentDisplay;
     }
 
-    public function displayThreadList($threads, $filtering, &$activeThreadAnnouncement, &$activeThreadTitle, &$activeThread, $thread_id_p, $current_categories_ids, $render, $is_full_page = false) {
-        $used_active = false; //used for the first one if there is not thread_id set
+    /**
+     * Renders or formats a list of threads.
+     * @param Thread[] $threads
+     * @param bool $render
+     * @param bool $is_full_page
+     * @return mixed[]|string
+     */
+    public function displayThreadList(array $threads, bool $render, bool $is_full_page = false): array|string {
         $current_user = $this->core->getUser()->getId();
-
-        $activeThreadAnnouncement = false;
-        $activeThreadTitle = "";
-        $activeThread = [];
-
         $thread_content = [];
 
-        $is_instructor_full_access = [];
-        $author_user_ids = array_map(function ($thread) {
-            return $thread["created_by"];
-        }, $threads);
-        $author_user_groups = $this->core->getQueries()->getAuthorUserGroups($author_user_ids);
-
-        foreach ($author_user_groups as $author) {
-            $is_instructor_full_access[$author["user_id"]] = $author["user_group"] <= User::GROUP_FULL_ACCESS_GRADER;
-        }
-
-        $thread_ids = array_column($threads, 'id');
-        $first_posts = $this->core->getQueries()->getFirstPostForThreads($thread_ids);
-        $viewed_threads = $this->core->getQueries()->getViewedThreads($current_user, $thread_ids);
-        $authors = $this->core->getQueries()->getUsersById($author_user_ids);
-        $authors_info = $this->core->getQueries()->getDisplayUserInfoFromUserIds($author_user_ids);
-
         foreach ($threads as $thread) {
-            // Checks if thread ID is empty. If so, skip this threads.
-            if (empty($thread["id"])) {
-                continue;
-            }
-            $first_post = $first_posts[$thread['id']] ?? null;
-            if (is_null($first_post)) {
+            $first_post = $thread->getFirstPost();
+            if ($first_post === false) {
                 // Thread without any posts(eg. Merged Thread)
-                $first_post = ['content' => "", 'render_markdown' => 0];
+                $first_post = new Post($thread);
                 $date = null;
             }
             else {
-                $date = DateUtils::convertTimeStamp($this->core->getUser(), $first_post['timestamp'], $this->core->getConfig()->getDateTimeFormat()->getFormat('forum'));
+                $date = DateUtils::convertTimeStamp($this->core->getUser(), DateUtils::dateTimeToString($first_post->getTimestamp()), $this->core->getConfig()->getDateTimeFormat()->getFormat('forum'));
             }
-            if ($thread['merged_thread_id'] != -1) {
+            if ($thread->isMergedThread()) {
                 // For the merged threads
-                $thread['status'] = 0;
+                $thread->setStatus(0);
             }
 
             $class = $is_full_page ? "thread_box thread-box-full" : "thread_box";
-            // $current_categories_ids should be subset of $thread["categories_ids"]
-            $issubset = (count(array_intersect($current_categories_ids, $thread["categories_ids"])) == count($current_categories_ids));
-            if (((isset($_REQUEST["thread_id"]) && $_REQUEST["thread_id"] == $thread["id"]) || $thread_id_p == $thread["id"] || $thread_id_p == -1) && !$used_active && $issubset) {
+            if (($_REQUEST["thread_id"] ?? -1) === $thread->getId()) {
                 $class .= " active";
-                $used_active = true;
-                $activeThreadTitle = "({$thread['id']}) " . $thread["title"];
-                $activeThread = $thread;
-                if ($thread["pinned_expiration"] > date("Y-m-d H:i:s")) {
-                    $activeThreadAnnouncement = true;
-                }
-                if ($thread_id_p == -1) {
-                    $thread_id_p = $thread["id"];
-                }
             }
-            $isNewThread = !in_array($thread['id'], $viewed_threads, true);
+            $isNewThread = $thread->isUnread($current_user);
             if ($isNewThread) {
                 $class .= " new_thread";
             }
-            if ($thread["deleted"]) {
+            if ($thread->isDeleted()) {
                 if ($isNewThread) {
                     $class .= " deleted-unviewed";
                 }
                 $class .= " deleted";
             }
 
-            if ($authors[$thread['created_by']]->accessGrading()) {
+            if ($thread->getAuthor()->accessGrading()) {
                 $class .= " important";
             }
 
-            //fix legacy code
-            $titleDisplay = $thread['title'];
-
             //replace tags from displaying in sidebar
-            $first_post_content = str_replace("`", "", strip_tags($first_post["content"]));
+            $first_post_content = str_replace("`", "", strip_tags($first_post->getContent()));
             $first_post_content = str_replace("#", "", $first_post_content);
             $temp_first_post_content = preg_replace('#\[(.*?)\]\((.*?)\)#', '$2', $first_post_content);
 
@@ -859,49 +653,32 @@ class ForumThreadView extends AbstractView {
                 $first_post_content = $temp_first_post_content;
             }
 
-            if ($first_post['render_markdown'] == 1) {
+            if ($first_post->isRenderMarkdown()) {
                 $first_post_content = $this->contentMarkdownToPlain($first_post_content);
             }
 
-            $sizeOfContent = strlen($first_post_content);
-            $titleLength = strlen($thread['title']);
+            $titleDisplay = $this->sizeTitle($thread->getTitle(), strlen($thread->getTitle()), $is_full_page ? 140 : 40);
+            $contentDisplay = $this->sizeContent(strlen($first_post_content), $first_post_content, $is_full_page ? 500 : 80);
 
-            if ($is_full_page) {
-                $titleDisplay = $this->sizeTitle($titleDisplay, $thread['title'], $titleLength, 140);
-                $contentDisplay = $this->sizeContent($sizeOfContent, $first_post_content, 500);
-            }
-            else {
-                $titleDisplay = $this->sizeTitle($titleDisplay, $thread['title'], $titleLength);
-                $contentDisplay = $this->sizeContent($sizeOfContent, $first_post_content);
-            }
+            $titleDisplay = "({$thread->getId()}) " . $titleDisplay;
 
-            $titleDisplay = "({$thread['id']}) " . $titleDisplay;
+            $link = $this->core->buildCourseUrl(['forum', 'threads', $thread->getId()]);
 
-            $link = $this->core->buildCourseUrl(['forum', 'threads', $thread['id']]);
-
-            $favorite = isset($thread['favorite']) && $thread['favorite'];
+            $favorite = $thread->isFavorite($current_user);
 
             $fa_icon = "fa-question";
             $fa_class = "thread-unresolved";
             $tooltip = "Thread Unresolved";
 
-            if (!isset($thread['status'])) {
-                $thread['status'] = 0;
-            }
-            if ($thread['status'] != 0) {
-                if ($thread['status'] == 1) {
-                    $fa_icon = "fa-check";
-                    $fa_class = "thread-resolved";
-                    $tooltip = "Thread Resolved";
-                }
+            if ($thread->getStatus() === 1) {
+                $fa_icon = "fa-check";
+                $fa_class = "thread-resolved";
+                $tooltip = "Thread Resolved";
             }
 
             $categories_content = [];
-            foreach ($thread["categories_desc"] as $category_desc) {
-                $categories_content[] = [$category_desc];
-            }
-            for ($i = 0; $i < count($thread["categories_color"]); $i += 1) {
-                $categories_content[$i][] = $thread["categories_color"][$i];
+            foreach ($thread->getCategories() as $category) {
+                $categories_content[] = [$category->getDescription(), $category->getColor()];
             }
 
             $date_content = ["not_null" => !is_null($date)];
@@ -910,27 +687,27 @@ class ForumThreadView extends AbstractView {
                 $date_content["formatted"] = $date;
             }
             $thread_info = [
-                'thread_id' => $thread['id'],
+                'thread_id' => $thread->getId(),
                 "title" => $titleDisplay,
                 "content" => $contentDisplay,
                 "categories" => $categories_content,
                 "link" => $link,
                 "class" => $class,
-                "pinned" => $thread["pinned_expiration"] > date("Y-m-d H:i:s"),
-                "expiring" => $thread["pinned_expiration"] <= date("Y-m-d H:i:s", strtotime("+7 day")),
+                "pinned" => $thread->isPinned(),
+                "expiring" => $thread->isPinnedExpiring(),
                 "favorite" => $favorite,
-                "merged_thread_id" => $thread['merged_thread_id'],
-                "status" => $thread["status"],
+                "merged_thread_id" => $thread->getMergedThread()?->getId() ?? -1,
+                "status" => $thread->getStatus(),
                 "fa_icon" => $fa_icon,
                 "fa_class" => $fa_class,
                 "tooltip" => $tooltip,
-                "is_locked" => isset($thread['lock_thread_date']) && $thread['lock_thread_date'] < date("Y-m-d H:i:S"),
+                "is_locked" => $thread->isLocked(),
                 "date" => $date_content,
-                "current_user_posted" => $thread["current_user_posted"]
+                "current_user_posted" => $thread->getAuthor()->getId() === $current_user,
             ];
 
             if ($is_full_page) {
-                $user_info = $authors_info[$first_post["author_user_id"]];
+                $user_info = $first_post->getAuthor()->getDisplayInfo();
                 $email = trim($user_info["user_email"]);
                 $given_name = trim($user_info["given_name"]);
                 $family_name = trim($user_info["family_name"]);
@@ -938,30 +715,30 @@ class ForumThreadView extends AbstractView {
                 $pronouns = trim($user_info["pronouns"]);
                 $display_pronouns = $user_info["display_pronouns"];
 
-                if ($is_instructor_full_access[$first_post["author_user_id"]]) {
+                if ($first_post->getAuthor()->accessFullGrading()) {
                     $visible_username = $given_name . " " . $family_name;
                 }
 
                 $author_info = [
-                    "user_id" => $first_post['author_user_id'],
-                    "name" => $first_post['anonymous'] ? "Anonymous" : $visible_username,
+                    "user_id" => $first_post->getAuthor()->getId(),
+                    "name" => $first_post->isAnonymous() ? "Anonymous" : $visible_username,
                     "email" => $email,
-                    "full_name" => $given_name . " " . $family_name . " (" . $first_post['author_user_id'] . ")",
+                    "full_name" => $given_name . " " . $family_name . " (" . $first_post->getAuthor()->getId() . ")",
                     "pronouns" => $pronouns,
                     "display_pronouns" => $display_pronouns
                 ];
                 $thread_info = array_merge($thread_info, [
-                    "post_id" => $first_post["id"],
-                    "is_thread_locked" => isset($thread['lock_thread_date']) && $thread['lock_thread_date'] < date("Y-m-d H:i:S"),
-                    "thread_resolve_state" => $thread['status'],
+                    "post_id" => $first_post->getId(),
+                    "is_thread_locked" => $thread->isLocked(),
+                    "thread_resolve_state" => $thread->getStatus(),
                     "show_unresolve" => false,
-                    "is_anon" => $first_post["anonymous"],
-                    "render_markdown" => $first_post["render_markdown"],
+                    "is_anon" => $first_post->isAnonymous(),
+                    "render_markdown" => $first_post->isRenderMarkdown(),
                     "author_info" => $author_info,
-                    "deleted" => $first_post['deleted']
+                    "deleted" => $first_post->isDeleted(),
+                    "sum_ducks" => $thread->getSumUpducks()
                 ]);
             }
-
             $thread_content[] = $thread_info;
         }
 
@@ -970,6 +747,7 @@ class ForumThreadView extends AbstractView {
         if ($render) {
             $return = $this->core->getOutput()->renderTwigTemplate("forum/displayThreadList.twig", [
                 "thread_content" => $thread_content,
+                "is_full_page" => $is_full_page,
             ]);
         }
         else {
@@ -1023,80 +801,30 @@ class ForumThreadView extends AbstractView {
     }
 
     /**
-     * @param array{
-     *      id: int,
-     *      title: string,
-     *      created_by: string,
-     *      pinned: bool,
-     *      deleted: bool,
-     *      merged_thread_id: int,
-     *      merged_post_id: int,
-     *      is_visible: bool,
-     *      status: int,
-     *      lock_thread_date?: string,
-     *      pinned_expiration: string,
-     *      announced: string
-     * } $thread
-     * @param array{
-     *      id: int,
-     *      thread_id: int,
-     *      parent_id: int,
-     *      author_user_id: string,
-     *      author_user_group?: string,
-     *      content: string,
-     *      timestamp: string,
-     *      edit_timestamp?: string,
-     *      anonymous: bool,
-     *      deleted: bool,
-     *      has_attachment: bool,
-     *      render_markdown: bool
-     * } $post
-     * @param array{
-     *      given_name: string,
-     *      family_name: string,
-     *      user_email: string,
-     *      pronouns: string,
-     *      display_pronouns: bool,
-     *      is_staff: bool
-     * } $author_info
-     * @param string[] $post_attachments
+     * Renders or formats a single post.
+     * @param Post $first_post
+     * @param Thread $thread
+     * @param Post $post
+     * @param bool $first
+     * @param string $display_option
+     * @param bool $includeReply
+     * @param int $post_box_id
+     * @param bool $render
+     * @return mixed[]|string
      */
-    public function createPost(string $first_post_author_id, bool $first_post_anonymous, array $thread, array $post, $unviewed_posts, $first, $reply_level, $display_option, int $counter, bool $isLiked, bool $taTrue, $includeReply, array $author_info, array $post_attachments, bool $has_history, bool $is_merged_thread, bool $render = false, bool $thread_announced = false, bool $isCurrentFavorite = false) {
-        $current_user = $this->core->getUser()->getId();
-        $thread_id = $thread["id"];
-        $post_id = $post["id"];
-        $parent_id = $post["parent_id"];
-
-        $thread_dir = FileUtils::joinPaths(FileUtils::joinPaths($this->core->getConfig()->getCoursePath(), "forum_attachments"), $thread_id);
-
+    public function createPost(Post $first_post, Thread $thread, Post $post, bool $first, string $display_option, bool $includeReply, int $post_box_id, bool $render = false): array|string {
+        $user = $this->core->getUser();
         // Get formatted time stamps
-        $date = DateUtils::convertTimeStamp($this->core->getUser(), $post['timestamp'], $this->core->getConfig()->getDateTimeFormat()->getFormat('forum'));
+        $date = DateUtils::convertTimeStamp($this->core->getUser(), DateUtils::dateTimeToString($post->getTimestamp()), $this->core->getConfig()->getDateTimeFormat()->getFormat('forum'));
 
-        if (isset($post["edit_timestamp"])) {
-            $edit_date = DateUtils::convertTimeStamp($this->core->getUser(), $post["edit_timestamp"], $this->core->getConfig()->getDateTimeFormat()->getFormat('forum'));
+        if (!$post->getHistory()->isEmpty()) {
+            $edit_timestamp = max($post->getHistory()->map(function ($x) {
+                return $x->getEditTimestamp();
+            })->toArray());
+            $edit_date = DateUtils::convertTimeStamp($this->core->getUser(), DateUtils::dateTimeToString($edit_timestamp), $this->core->getConfig()->getDateTimeFormat()->getFormat('forum'));
         }
         else {
             $edit_date = null;
-        }
-
-        $author_email = trim($author_info['user_email']);
-        $given_name = trim($author_info["given_name"]);
-        $family_name = trim($author_info["family_name"]);
-        $visible_username = $given_name . " " . substr($family_name, 0, 1) . ".";
-        $pronouns = trim($author_info["pronouns"]);
-        $display_pronouns = $author_info["display_pronouns"];
-        $thread_resolve_state = $thread['status'];
-
-        if ($display_option != 'tree') {
-            $reply_level = 1;
-        }
-
-        if (isset($post["author_user_group"]) && $post["author_user_group"] <= User::GROUP_FULL_ACCESS_GRADER) {
-            $visible_username = $given_name . " " . $family_name;
-        }
-
-        if ($post["anonymous"]) {
-            $visible_username = "Anonymous";
         }
 
         $classes = ["post_box"];
@@ -1104,28 +832,21 @@ class ForumThreadView extends AbstractView {
             $classes[] = "first_post";
         }
         $isNewPost = false;
-        if (in_array($post_id, $unviewed_posts, true)) {
-            if ($current_user != $post["author_user_id"]) {
-                $classes[] = "new_post";
-                $isNewPost = true;
-            }
-            elseif ($current_user === $post["author_user_id"]) {
-                $classes[] = "new_post";
-                $isNewPost = true;
+        if ($thread->getNewPosts($user->getId())->contains($post)) {
+            $classes[] = "new_post";
+            $isNewPost = true;
+            if ($post->getAuthor()->accessGrading()) {
+                $classes[] = "important important-new";
             }
         }
         else {
             $classes[] = "viewed_post";
-        }
-        if ($author_info['is_staff']) {
-            if (in_array($post_id, $unviewed_posts, true)) {
-                $classes[] = "important new_post important-new";
-            }
-            else {
+            if ($post->getAuthor()->accessGrading()) {
                 $classes[] = "important";
             }
         }
-        if ($post["deleted"]) {
+
+        if ($post->isDeleted()) {
             $classes[] = "deleted";
             if ($isNewPost) {
                 $classes[] = "deleted-unviewed";
@@ -1136,118 +857,113 @@ class ForumThreadView extends AbstractView {
             $deleted = false;
         }
 
-        $offset = min(($reply_level - 1) * 30, 180);
+        $offset = min(($post->getReplyLevel() - 1) * 30, 180);
 
-        $post_content = $post['content'];
-        $markdown = $post["render_markdown"];
-
-        $isThreadLocked = isset($thread['lock_thread_date']) && $thread['lock_thread_date'] < date("Y-m-d H:i:S");
-        $userAccessFullGrading = $this->core->getUser()->accessFullGrading();
-        $userGroup = $this->core->getUser()->getGroup();
+        $post_content = $post->getContent();
+        do {
+            $post_content = preg_replace('/(?:!\[(.*?)\]\((.*?)\))/', '$2', $post_content, -1, $count);
+        } while ($count > 0);
 
         $post_user_info = [];
 
-        $merged_thread = $is_merged_thread && $userAccessFullGrading;
+        $merged_thread = $thread->isMergedThread() && $user->accessFullGrading();
 
         $post_button = [];
 
-        if ($userGroup <= User::GROUP_LIMITED_ACCESS_GRADER || $post['author_user_id'] === $current_user) {
-            if ($isThreadLocked && !$userAccessFullGrading) {
+        if (($user->accessGrading() || $post->getAuthor()->getId() === $user->getId()) && (!$thread->isLocked() || $user->accessFullGrading())) {
+            if ($deleted && $user->accessGrading()) {
+                $ud_toggle_status = "false";
+                $ud_button_title = "Undelete post";
+                $ud_button_icon = "fa-undo";
             }
             else {
-                if ($deleted && $userGroup <= User::GROUP_LIMITED_ACCESS_GRADER) {
-                    $ud_toggle_status = "false";
-                    $ud_button_title = "Undelete post";
-                    $ud_button_icon = "fa-undo";
-                }
-                else {
-                    $ud_toggle_status = "true";
-                    $ud_button_title = "Remove post";
-                    $ud_button_icon = "fa-trash";
-                }
-
-                $post_button["delete"] = [
-                    "ud_toggle_status" => $ud_toggle_status,
-                    "csrf_token" => $this->core->getCsrfToken(),
-                    "ud_button_title" => $ud_button_title,
-                    "ud_button_icon" => $ud_button_icon
-                ];
-
-                $shouldEditThread = null;
-
-                if ($first) {
-                    $shouldEditThread = "true";
-                    $edit_button_title = "Edit thread and post";
-                }
-                else {
-                    $shouldEditThread = "false";
-                    $edit_button_title = "Edit post";
-                }
-
-                $post_button["edit"] = [
-                    "shouldEditThread" => $shouldEditThread,
-                    "edit_button_title" => $edit_button_title,
-                    "csrf_token" => $this->core->getCsrfToken()
-                ];
+                $ud_toggle_status = "true";
+                $ud_button_title = "Remove post";
+                $ud_button_icon = "fa-trash";
             }
-        }
 
-        $post_up_duck = [
-            "upduck_count" => $counter,
-            "upduck_user_liked" => $isLiked,
-            "taTrue" => $taTrue
-        ];
+            $post_button["delete"] = [
+                "ud_toggle_status" => $ud_toggle_status,
+                "csrf_token" => $this->core->getCsrfToken(),
+                "ud_button_title" => $ud_button_title,
+                "ud_button_icon" => $ud_button_icon
+            ];
 
-        if ($userGroup <= User::GROUP_STUDENT) {
-            $info_name = $given_name . " " . $family_name . " (" . $post['author_user_id'] . ")";
-            $visible_user_json = json_encode($visible_username);
-            $pronouns = trim($author_info["pronouns"]);
-            $display_pronouns = $author_info["display_pronouns"];
-            $info_name = json_encode($info_name);
-            $jscriptAnonFix = $post['anonymous'] ? 'true' : 'false';
-            $jscriptAnonFix = json_encode($jscriptAnonFix);
+            if ($first) {
+                $shouldEditThread = "true";
+                $edit_button_title = "Edit thread and post";
+            }
+            else {
+                $shouldEditThread = "false";
+                $edit_button_title = "Edit post";
+            }
 
-            $post_user_info = [
-                "info_name" => $info_name,
-                "visible_user_json" => $visible_user_json,
-                "jscriptAnonFix" => $jscriptAnonFix,
-                "pronouns" => $pronouns,
-                "display_pronouns" => $display_pronouns
+            $post_button["edit"] = [
+                "shouldEditThread" => $shouldEditThread,
+                "edit_button_title" => $edit_button_title,
+                "csrf_token" => $this->core->getCsrfToken()
             ];
         }
 
-        $post_user_info["is_OP"] = ($post["author_user_id"] === $first_post_author_id) && ($first_post_anonymous === $post["anonymous"]);
+        $post_up_duck = [
+            "upduck_count" => count($post->getUpduckers()),
+            "upduck_user_liked" => $post->getUpduckers()->map(function ($x) {
+                return $x->getId();
+            })->contains($user->getId()),
+            "taTrue" => !$post->getUpduckers()->filter(function ($x) {
+                return $x->accessFullGrading();
+            })->isEmpty(),
+        ];
+
+        $author_display_info = $post->getAuthor()->getDisplayInfo();
+        $visible_username = $author_display_info["given_name"] . " " . substr($author_display_info["family_name"], 0, 1) . ".";
+
+        if ($post->getAuthor()->accessFullGrading()) {
+            $visible_username = $author_display_info["given_name"] . " " . $author_display_info["family_name"];
+        }
+
+        if ($post->isAnonymous()) {
+            $visible_username = "Anonymous";
+        }
+
+        $post_user_info = [
+            "info_name" => json_encode($author_display_info["given_name"] . " " . $author_display_info["family_name"] . " (" . $post->getAuthor()->getId() . ")"),
+            "visible_user_json" => json_encode($visible_username),
+            "jscriptAnonFix" => json_encode($post->isAnonymous() ? 'true' : 'false'),
+            "pronouns" => trim($author_display_info["pronouns"]),
+            "display_pronouns" => $author_display_info["display_pronouns"],
+            "is_OP" => ($post->getAuthor()->getId() === $first_post->getAuthor()->getId()) && ($post->isAnonymous() === $first_post->isAnonymous()),
+        ];
 
         $post_attachment = ForumUtils::getForumAttachments(
-            $post_id,
-            $thread_id,
-            $post_attachments,
+            $post->getId(),
+            $thread->getId(),
+            $post->getAttachments()->filter(function ($x) {
+                return $x->isCurrent();
+            })->map(function ($x) {
+                return $x->getFileName();
+            })->toArray(),
             $this->core->getConfig()->getCoursePath(),
             $this->core->buildCourseUrl(['display_file'])
         );
 
-        $post_box_id = 1;
-        if (!$isThreadLocked || $this->core->getUser()->accessFullGrading()) {
-            $GLOBALS['post_box_id'] = $post_box_id = isset($GLOBALS['post_box_id']) ? $GLOBALS['post_box_id'] + 1 : 2;
-        }
-
         $created_post = [
             "classes" => $classes,
-            "post_id" => $post_id,
-            "reply_level" => $reply_level,
+            "post_id" => $post->getId(),
+            "reply_level" => $post->getReplyLevel(),
             "offset" => $offset,
             "first" => $first,
             "post_content" => $post_content,
             "post" => $post,
             "display_option" => $display_option,
-            "isThreadLocked" => $isThreadLocked,
-            "userAccessFullGrading" => $userAccessFullGrading,
-            "userGroup" => $userGroup,
+            "isThreadLocked" => $thread->isLocked(),
+            "userAccessFullGrading" => $user->accessFullGrading(),
+            "userGroup" => $user->getGroup(),
             "includeReply" => $includeReply,
-            "thread_resolve_state" => $thread_resolve_state,
+            "thread_resolve_state" => $thread->getStatus(),
             "show_unresolve" => false,
-            "current_user" => $current_user,
-            "author_email" => $author_email,
+            "current_user" => $user->getId(),
+            "author_email" => $author_display_info["user_email"],
             "post_user_info" => $post_user_info,
             "post_up_duck" => $post_up_duck,
             "post_date" => $date,
@@ -1257,25 +973,21 @@ class ForumThreadView extends AbstractView {
             "post_attachment" => $post_attachment,
             "form_post_url" => $this->core->buildCourseUrl(['forum', 'posts', 'new']),
             "post_box_id" => $post_box_id,
-            "thread_id" => $thread_id,
-            "parent_id" => $parent_id,
-            "render_markdown" => $markdown,
-            "has_history" => $has_history,
+            "thread_id" => $thread->getId(),
+            "parent_id" => $post->getParent()->getId(),
+            "render_markdown" => $post->isRenderMarkdown(),
+            "has_history" => !$post->getHistory()->isEmpty(),
             "thread_previously_merged" => $merged_thread,
-            "thread_announced" => $thread_announced,
+            "thread_announced" => $thread->isAnnounced(),
         ];
 
         if ($render) {
             if ($first) {
-                $thread_title = $thread['title'];
-                $activeThreadTitle = "({$thread_id}) " . $thread_title;
-
-                $created_post['activeThreadTitle'] = $activeThreadTitle;
-                $activeThreadAnnouncement = $thread['pinned_expiration'] > date("Y-m-d H:i:s");
-                $created_post['activeThreadAnnouncement'] = $activeThreadAnnouncement;
+                $created_post['activeThreadTitle'] = "({$thread->getId()}) " . $thread->getTitle();
+                $created_post['activeThreadAnnouncement'] = $thread->isPinned();
             }
             $created_post['activeThread'] = $thread;
-            $created_post['isCurrentFavorite'] = $isCurrentFavorite;
+            $created_post['isCurrentFavorite'] = $thread->isFavorite($user->getId());
             $created_post['csrf_token'] = $this->core->getCsrfToken();
             return $this->core->getOutput()->renderTwigTemplate("forum/CreatePost.twig", $created_post);
         }
