@@ -14,6 +14,7 @@ use app\models\gradeable\GradeInquiry;
 use app\models\SimpleStat;
 use app\models\Team;
 use app\models\User;
+use app\entities\forum\Thread;
 use app\views\AbstractView;
 use app\libraries\NumberUtils;
 use app\libraries\CodeMirrorUtils;
@@ -25,6 +26,7 @@ class ElectronicGraderView extends AbstractView {
      * @param Gradeable $gradeable
      * @param array[] $sections
      * @param SimpleStat[] $component_averages
+     * @param SimpleStat|null $manual_average
      * @param SimpleStat|null $autograded_average
      * @param SimpleStat|null $overall_average
      * @param int $total_submissions
@@ -42,6 +44,7 @@ class ElectronicGraderView extends AbstractView {
         Gradeable $gradeable,
         array $sections,
         array $component_averages,
+        $manual_average,
         $autograded_average,
         $overall_scores,
         $overall_average,
@@ -273,6 +276,7 @@ class ElectronicGraderView extends AbstractView {
             if (count($component_averages) !== 0) {
                 foreach ($component_averages as $comp) {
                     /* @var SimpleStat $comp */
+                    $manual_average += $comp->getAverageScore();
                     $component_overall_score += $comp->getAverageScore();
                     $component_overall_max += $comp->getMaxValue();
                     $percentage = 0;
@@ -348,6 +352,7 @@ class ElectronicGraderView extends AbstractView {
             "overall_percentage" => $overall_percentage,
             "autograded_percentage" => $autograded_percentage,
             "autograded_average" => $autograded_average,
+            "manual_average" => $manual_average,
             "component_averages" => $component_averages,
             "component_percentages" => $component_percentages,
             "component_overall_score" => $component_overall_score,
@@ -1343,8 +1348,6 @@ HTML;
             <span class="col grading_label">Discussion Posts</span>
 HTML;
 
-        $currentCourse = $this->core->getConfig()->getCourse();
-
         //Empty thread input
         if ($threadIds === "{}") {
             $threadIds = [];
@@ -1359,27 +1362,24 @@ HTML;
             $id = $submitter->getId();
             $submitters = [$id];
         }
-        foreach ($threadIds as $threadId) {
-            $posts = [];
-            foreach ($submitters as $s_id) {
-                $posts = array_merge($posts, $this->core->getQueries()->getPostsForThread($this->core->getUser()->getId(), $threadId, false, 'time', $s_id));
-            }
-            if (count($posts) > 0) {
-                $posts_view .= $this->core->getOutput()->renderTemplate('forum\ForumThread', 'generatePostList', $threadId, $posts, [], $currentCourse, false, true, $id);
+        $repo = $this->core->getCourseEntityManager()->getRepository(Thread::class);
+        $threads = $repo->getThreadsForGrading($threadIds, $submitters);
+        foreach ($threads as $thread) {
+            if (count($thread->getPosts()) > 0) {
+                $posts_view .= $this->core->getOutput()->renderTemplate('forum\ForumThread', 'generatePostList', $thread, false, "alpha", []);
             }
             else {
                 $posts_view .= <<<HTML
-                    <h3 style="text-align: center;">No posts for thread id: {$threadId}</h3> <br/>
+                <h3 style="text-align: center;">No posts for thread id: {$thread->getId()}</h3> <br/>
 HTML;
             }
-
             $posts_view .= <<<HTML
-                    <a href="{$this->core->buildCourseUrl(['forum', 'threads', $threadId])}" target="_blank" rel="noopener nofollow" class="btn btn-default btn-sm" style="margin-top:15px; text-decoration: none;" onClick="" data-testid="go-to-thread"> Go to thread</a>
+                    <a href="{$this->core->buildCourseUrl(['forum', 'threads', $thread->getId()])}" target="_blank" rel="noopener nofollow" class="btn btn-default btn-sm" style="margin-top:15px; text-decoration: none;" onClick="" data-testid="go-to-thread"> Go to thread</a>
                     <hr style="border-top:1px solid #999;margin-bottom: 5px;" /> <br/>
 HTML;
         }
 
-        if (empty($threadIds)) {
+        if (count($threads) === 0) {
             $posts_view .= <<<HTML
                 <h3 style="text-align: center;">No thread id specified.</h3> <br/>
 HTML;
