@@ -2,52 +2,103 @@ import { isPermissionAllowed } from 'cypress-browser-permissions';
 
 const queueName = 'Cypress Office Hour Queue 1';
 const queueName_random = 'Cypress Office Hour Queue Random';
-const queueName1 = 'Cypress Office Hour Queue 2';
+const queueName_blank = 'Cypress Office Hour Queue Blank';
 const queueCode = 'cypress_test';
 const queueCode1 = 'cypress_test_fail';
 const newQueueCode = 'cypress_update';
 
-const enableQueue = () => {
-    cy.visit(['sample', 'config']); // course setting
-    cy.get('[data-testid="queue-enabled"]').check();
-    cy.get('[data-testid="queue-enabled"]').should('be.checked');
+const bitdiddleRows = [{
+    state: 'done',
+    queue: queueName,
+    helpedBy: '-',
+    removedBy: 'bitdiddle',
+    removalMethod: 'self',
+},
+{
+    state: 'done',
+    queue: queueName_blank,
+    helpedBy: '-',
+    removedBy: 'bitdiddle',
+    removalMethod: 'self',
+},
+{
+    state: 'done',
+    queue: queueName,
+    helpedBy: 'instructor',
+    removedBy: 'bitdiddle',
+    removalMethod: 'self_helped',
+},
+{
+    state: 'done',
+    queue: queueName,
+    helpedBy: '-',
+    removedBy: 'instructor',
+    removalMethod: 'emptied',
+}];
+
+const wisozaRows = [{
+    state: 'waiting',
+    queue: queueName_blank,
+    helpedBy: '-',
+    removedBy: '-',
+    removalMethod: '-',
+}];
+
+const checkRows = (rows) => {
+    cy.get('[data-testid="row-label"]')
+        .its('length')
+        .then((length) => {
+            const startingRow = length - rows.length + 1;
+            for (let i = 0; i < rows.length; i++) {
+                cy.get(`[data-testid="student-row-${startingRow + i}"]`).first().as(`row-${i}`);
+                cy.get(`@row-${i}`).find('[data-testid="row-label"]').should('contain', startingRow + i);
+                cy.get(`@row-${i}`).find('[data-testid="current-state"]').should('contain', rows[i].state);
+                cy.get(`@row-${i}`).find('[data-testid="queue"]').should('contain', rows[i].queue);
+                if (rows[i].state !== 'waiting') {
+                    // This checks if time entered and time removed are in fact times.
+                    // We do not check for a specific time because this may change.
+                    cy.get(`@row-${i}`).find('[data-testid="time-entered"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
+                    cy.get(`@row-${i}`).find('[data-testid="time-removed"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
+                }
+                cy.get(`@row-${i}`).find('[data-testid="helped-by"]').should('contain', rows[i].helpedBy);
+                cy.get(`@row-${i}`).find('[data-testid="removed-by"]').should('contain', rows[i].removedBy);
+                cy.get(`@row-${i}`).find('[data-testid="removal-method"]').should('contain', rows[i].removalMethod);
+            }
+        });
 };
 
-const deleteQueue = () => {
+const deleteQueue = (queue_name) => {
     cy.visit(['sample', 'office_hours_queue']); // office hours queue
     cy.get('[data-testid="toggle-filter-settings"]').click();
     cy.get('[data-testid="popup-window"]').should('exist');
-    cy.get('[data-testid="delete-queue-btn"]').last().click();
+    cy.get('[data-testid="queue-name"]').contains(queue_name).parents('[data-testid="queue-item"]').find('[data-testid="delete-queue-btn"]').click();
 };
 
-const disableQueue = () => {
-    cy.visit(['sample', 'config']);
-    cy.get('[data-testid="queue-enabled"]').should('be.checked');
-    cy.get('[data-testid="queue-enabled"]').uncheck();
-};
 const openNewQueue = (queueName, queueCode = '') => {
-    cy.get('#nav-sidebar-queue').click();
+    cy.visit(['sample', 'office_hours_queue']);
     cy.get('[data-testid="toggle-new-queue"]').click();
     cy.get('[data-testid="popup-window"]').should('exist');
     cy.get('[data-testid="new-queue-code"]').type(queueName);
-    if (queueCode.length > 0) {
-        cy.get('[data-testid="new-queue-token"]').type(queueCode);
+    if (queueCode === 'RANDOM') {
+        cy.get('[data-testid="new-queue-rand-token"]').click(); // random code
+        cy.get('[data-testid="new-queue-token"]').invoke('val').should('not.be.empty');
     }
-    else {
-        cy.get('[data-testid="new-queue-rand-token"]').click();
+    else if (queueCode !== '') {
+        cy.get('[data-testid="new-queue-token"]').type(queueCode);
     }
     cy.get('[data-testid="open-new-queue-btn"]').click();
 };
 
-const changeQueueCode = (queueName, queueCode = '') => {
+const changeQueueCode = (queueName, queueCode) => {
     cy.get('[data-testid="toggle-filter-settings"]').click();
     cy.get('[data-testid="popup-window"]').should('exist');
     cy.get('[data-testid="old-queue-code"]').select(queueName);
-    if (queueCode.length > 0) {
-        cy.get('[data-testid="old-queue-token"]').type(queueCode);
-    }
-    else {
+    if (queueCode === 'RANDOM') {
         cy.get('[data-testid="old-queue-rand-token"]').click(); // random code
+        cy.get('[data-testid="old-queue-token"]').invoke('val').should('not.be.empty');
+    }
+    else if (queueCode !== '') {
+        cy.get('[data-testid="old-queue-token"]').type(queueCode);
     }
     cy.get('[data-testid="change-code-btn"]').click(); // update it
 };
@@ -62,9 +113,12 @@ const studentJoinQueue = (queueName, queueCode) => {
     cy.get('[data-testid="queue-code"]').select(queueName);
     cy.get('[data-testid="queue-code"]').invoke('val'); // in which queue you want to join
     cy.get('[data-testid="queue-code"]').should('contain', queueName);
-    cy.get('#token-box').type(queueCode);
+    if (queueCode !== '') {
+        cy.get('#token-box').type(queueCode);
+    }
     cy.get('[data-testid="join-queue-btn"]').click();
 };
+
 const editAnnouncement = (text = '') => {
     // openAnnouncementSettings
     cy.get('[data-testid="toggle-announcement-settings"]').click();
@@ -78,74 +132,117 @@ const editAnnouncement = (text = '') => {
 };
 
 describe('test office hours queue', () => {
-    it('Testing queue as student', () => {
+    before(() => {
         cy.login();
-        enableQueue();
-        // deleting the Lab help and homework debugging
-        deleteQueue();
-        deleteQueue();
+
+        // enable queue if not already enabled
+        cy.visit(['sample', 'config']); // course setting
+        cy.get('[data-testid="queue-enabled"]').check();
+        cy.get('[data-testid="queue-enabled"]').should('be.checked');
+        cy.reload();
+
         openNewQueue(queueName, queueCode);
-        openNewQueue(queueName, queueCode1); // same name but used different code
+        cy.get('[data-testid="popup-message"]').should('contain', 'New queue added');
+        openNewQueue(queueName_random, 'RANDOM');
+        cy.get('[data-testid="popup-message"]').should('contain', 'New queue added');
+        openNewQueue(queueName_blank, '');
+        cy.get('[data-testid="popup-message"]').should('contain', 'New queue added');
+    });
+
+    it('Creating queues and changing queue codes', () => {
+        cy.login();
+        // Create queue with same name but different code to ensure error
+        openNewQueue(queueName, queueCode1);
         cy.get('[data-testid="popup-message"]').should('contain', 'Unable to add queue. Make sure you have a unique queue name');
 
-        openNewQueue(queueName_random);
-        changeQueueCode(queueName_random);
+        changeQueueCode(queueName_random, 'RANDOM');
         cy.get('[data-testid="popup-message"]').should('contain', 'Queue Access Code Changed');
 
+        changeQueueCode(queueName_blank, '');
+        cy.get('[data-testid="popup-message"]').should('contain', 'Queue Access Code Changed');
+
+        // Change 1st queue's code
         changeQueueCode(queueName, newQueueCode);
         cy.get('[data-testid="popup-message"]').should('contain', 'Queue Access Code Changed');
+    });
 
-        openNewQueue(queueName1, queueCode1);
-
+    it('Joining queues as student', () => {
         // switch to student to join queue
-        switchUser('student');
-        // cy.get('#leave_queue').click();
+        switchUser('bitdiddle');
 
+        // Join first queue created with user 'bitdiddle'
         studentJoinQueue(queueName, newQueueCode);
         cy.get('[data-testid="popup-message"]').should('contain', 'Added to queue');
+        // Remove user 'bitdiddle' from queue
         cy.get('[data-testid="leave-queue"]').click(); // studentRemoveSelfFromQueue
         cy.get('[data-testid="popup-message"]').should('contain', 'Removed from queue');
 
+        // Join first queue created with blank code with user 'bitdiddle'
+        studentJoinQueue(queueName_blank, '');
+        cy.get('[data-testid="popup-message"]').should('contain', 'Added to queue');
+        // Remove user 'bitdiddle' from queue
+        cy.get('[data-testid="leave-queue"]').click(); // studentRemoveSelfFromQueue
+        cy.get('[data-testid="popup-message"]').should('contain', 'Removed from queue');
+
+        // Rejoin first queue created with user 'bitdiddle'
         studentJoinQueue(queueName, newQueueCode);
         cy.get('[data-testid="popup-message"]').should('contain', 'Added to queue');
-
+    });
+    it('Helping student', () => {
         // switch to instructor to help first student
         switchUser('instructor');
-        cy.get('.help_btn').first().click(); // helpFirstStudent
+        // help 'bitdiddle' in first queue created
+        cy.get('.help_btn').last().click();
         cy.get('[data-testid="popup-message"]').should('contain', 'Started helping student');
 
         // switch to student for finishing help
-        switchUser('student');
-        cy.get('[data-testid="self-finish-help"]').click(); // studentFinishHelpSelf
+        switchUser('bitdiddle');
+        cy.get('[data-testid="self-finish-help"]').click();
         cy.get('[data-testid="popup-message"]').should('contain', 'Finished helping student');
-        studentJoinQueue(queueName1, queueCode1);
+        // Rejoin queue as 'bitdiddle'
+        studentJoinQueue(queueName, newQueueCode);
         cy.get('[data-testid="popup-message"]').should('contain', 'Added to queue');
 
-        // switch to student (aphacker) for joining queue
-        switchUser('aphacker');
-        studentJoinQueue(queueName1, queueCode1);
+        // switch to student (wisoza)
+        switchUser('wisoza');
+        // Join blank queue with user 'wisoza'
+        studentJoinQueue(queueName_blank, '');
         cy.get('[data-testid="popup-message"]').should('contain', 'Added to queue');
     });
 
     it('Helping,removing student and toggle queue as Instructor', () => {
+        // Switch to instructor for helping students and changing settings
         switchUser('instructor');
-        cy.get('.help_btn').first().click(); // help first student
+        // Start helping 'bitdiddle' in first queue created
+        cy.get('.help_btn').eq(4).click();
         cy.get('[data-testid="popup-message"]').should('contain', 'Started helping student');
-        cy.get('.finish_helping_btn').first().click(); // finished helping first student
+        // Finish helping 'bitdiddle' in first queue created
+        cy.get('.finish_helping_btn').last().click(); // finished helping first student
         cy.get('[data-testid="popup-message"]').should('contain', 'Finished helping student');
-        cy.get('[data-testid="remove-from-queue-btn"]').first().click(); // remove First Student
+        // Remove 'wisoza' from queue
+        cy.get('[data-testid="remove-from-queue-btn"]').last().click(); // remove First Student
         cy.get('[data-testid="popup-message"]').should('contain', 'Removed from queue');
+        // Restore 'bitdiddle' to newly created queue
         cy.get('[data-testid="queue-restore-btn"]').first().click(); // restore first Student
         cy.get('[data-testid="popup-message"]').should('contain', 'Student restored');
+        // Restore 'wisoza' to blank queue
         cy.get('[data-testid="queue-restore-btn"]').first().click();
         cy.get('[data-testid="popup-message"]').should('contain', 'Student restored');
+        // Attempt to restore 'bitdiddle' to newly created queue
         cy.get('[data-testid="queue-restore-btn"]').first().click();
         cy.get('[data-testid="popup-message"]').should('contain', 'Cannot restore a user that is currently in the queue. Please remove them first.');
+        // Ensure Wally is shown in queue before filtered
+        cy.contains('huelsw@sample.edu').should('be.visible');
+        // Filter to show only 'Lab Help' queue
         cy.get('.filter-buttons').first().click(); // turn first "off"
+        // Ensure Wally isn't shown in queue after filtered
+        cy.contains('huelsw@sample.edu').should('not.be.visible');
         cy.get('.filter-buttons').first().click(); // turn first "on"
         cy.get('[data-testid="toggle-filter-settings"]').first().click();
-        cy.get('[data-testid="toggle-queue-checkbox"]').first().click(); // closeFirstQueue
-        cy.get('[data-testid="empty-queue-btn"]').first().click(); // emptyFirstQueue
+        // Close the newly created queue
+        cy.get('[data-testid="toggle-queue-checkbox"]').eq(2).click(); // closeFirstQueue
+        // Empty newly created queue
+        cy.get('[data-testid="empty-queue-btn"]').eq(2).click(); // emptyFirstQueue
 
         editAnnouncement('Submitty');
         cy.get('[data-testid="popup-message"]').should('contain', 'Updated announcement');
@@ -155,106 +252,36 @@ describe('test office hours queue', () => {
         cy.get('[data-testid="popup-message"]').should('contain', 'Updated announcement');
 
         // Confirm student queue history
-        cy.get('[data-testid="search-student-queue-input"]').first().type('student');
+        cy.get('[data-testid="search-student-queue-input"]').first().type('bitdiddle');
         cy.get('[data-testid="search-student-queue-btn"]').first().click();
+        cy.contains('(ID:bitdiddle)').should('be.visible');
 
-        cy.get('[data-testid="student-row-1"]').first().as('row-1');
-        cy.get('@row-1').find('[data-testid="row-label"]').should('contain', '1');
-        cy.get('@row-1').find('[data-testid="current-state"]').should('contain', 'done');
-        cy.get('@row-1').find('[data-testid="queue"]').should('contain', 'Lab Help');
-        // This checks if time entered and time removed are in fact times.
-        // We do not check for a specific time because this may change.
-        cy.get('@row-1').find('[data-testid="time-entered"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
-        cy.get('@row-1').find('[data-testid="time-removed"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
-        cy.get('@row-1').find('[data-testid="helped-by"]').should('contain', 'grader');
-        cy.get('@row-1').find('[data-testid="removed-by"]').should('contain', 'grader');
-        cy.get('@row-1').find('[data-testid="removal-method"]').should('contain', 'helped');
-
-        cy.get('[data-testid="student-row-2"]').first().as('row-2');
-        cy.get('@row-2').find('[data-testid="row-label"]').should('contain', '2');
-        cy.get('@row-2').find('[data-testid="current-state"]').should('contain', 'done');
-        cy.get('@row-2').find('[data-testid="queue"]').should('contain', 'Lab Help');
-        cy.get('@row-2').find('[data-testid="time-entered"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
-        cy.get('@row-2').find('[data-testid="time-removed"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
-        cy.get('@row-2').find('[data-testid="helped-by"]').should('contain', '-');
-        cy.get('@row-2').find('[data-testid="removed-by"]').should('contain', 'instructor');
-        cy.get('@row-2').find('[data-testid="removal-method"]').should('contain', 'emptied');
-
-        cy.get('[data-testid="student-row-3"]').first().as('row-3');
-        cy.get('@row-3').find('[data-testid="row-label"]').should('contain', '3');
-        cy.get('@row-3').find('[data-testid="current-state"]').should('contain', 'done');
-        cy.get('@row-3').find('[data-testid="queue"]').should('contain', queueName);
-        cy.get('@row-3').find('[data-testid="time-entered"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
-        cy.get('@row-3').find('[data-testid="time-removed"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
-        cy.get('@row-3').find('[data-testid="helped-by"]').should('contain', '-');
-        cy.get('@row-3').find('[data-testid="removed-by"]').should('contain', 'student');
-        cy.get('@row-3').find('[data-testid="removal-method"]').should('contain', 'self');
-
-        cy.get('[data-testid="student-row-4"]').first().as('row-4');
-        cy.get('@row-4').find('[data-testid="row-label"]').should('contain', '4');
-        cy.get('@row-4').find('[data-testid="current-state"]').should('contain', 'done');
-        cy.get('@row-4').find('[data-testid="queue"]').should('contain', queueName);
-        cy.get('@row-4').find('[data-testid="time-entered"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
-        cy.get('@row-4').find('[data-testid="time-removed"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
-        cy.get('@row-4').find('[data-testid="helped-by"]').should('contain', 'instructor');
-        cy.get('@row-4').find('[data-testid="removed-by"]').should('contain', 'student');
-        cy.get('@row-4').find('[data-testid="removal-method"]').should('contain', 'self_helped');
-
-        cy.get('[data-testid="student-row-5"]').first().as('row-5');
-        cy.get('@row-5').find('[data-testid="row-label"]').should('contain', '5');
-        cy.get('@row-5').find('[data-testid="current-state"]').should('contain', 'waiting');
-        cy.get('@row-5').find('[data-testid="queue"]').should('contain', queueName1);
-        cy.get('@row-5').find('[data-testid="time-entered"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
-        cy.get('@row-5').find('[data-testid="time-removed"]').should('contain', '-');
-        cy.get('@row-5').find('[data-testid="helped-by"]').should('contain', '-');
-        cy.get('@row-5').find('[data-testid="removed-by"]').should('contain', '-');
-        cy.get('@row-5').find('[data-testid="removal-method"]').should('contain', '-');
+        checkRows(bitdiddleRows);
 
         cy.get('#times-helped-cell').should('contain', '1 times helped.');
 
-        // Confirm aphacker queue history
+        // Confirm wisoza queue history
         // Use search autocomplete feature
         cy.get('[data-testid="search-student-queue-input"]').first().as('queue-search');
         cy.get('@queue-search').clear();
-        cy.get('@queue-search').type('hack');
+        cy.get('@queue-search').type('wisoza');
 
         cy.get('#ui-id-1').first().should('be.visible');
         cy.get('#ui-id-1').click();
-        cy.get('[data-testid="search-student-queue-input"]').first().should('have.value', 'aphacker');
+        cy.get('[data-testid="search-student-queue-input"]').first().should('have.value', 'wisoza');
         cy.get('[data-testid="search-student-queue-btn"]').first().click();
+        cy.contains('(ID:wisoza)').should('be.visible');
 
-        cy.get('[data-testid="student-row-1"]').first().as('row-1');
-        cy.get('@row-1').find('[data-testid="row-label"]').should('contain', '1');
-        cy.get('@row-1').find('[data-testid="current-state"]').should('contain', 'done');
-        cy.get('@row-1').find('[data-testid="queue"]').should('contain', 'Homework Debugging');
-        cy.get('@row-1').find('[data-testid="time-entered"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
-        cy.get('@row-1').find('[data-testid="time-removed"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
-        cy.get('@row-1').find('[data-testid="helped-by"]').should('contain', 'ta');
-        cy.get('@row-1').find('[data-testid="removed-by"]').should('contain', 'instructor');
-        cy.get('@row-1').find('[data-testid="removal-method"]').should('contain', 'emptied');
-
-        cy.get('[data-testid="student-row-2"]').first().as('row-2');
-        cy.get('@row-2').find('[data-testid="row-label"]').should('contain', '2');
-        cy.get('@row-2').find('[data-testid="current-state"]').should('contain', 'waiting');
-        cy.get('@row-2').find('[data-testid="queue"]').should('contain', 'Cypress Office Hour Queue 2');
-        cy.get('@row-2').find('[data-testid="time-entered"]').invoke('text').should('match', /\d{4}-\d\d-\d\d \d\d:\d\d:\d\d.*/);
-        cy.get('@row-2').find('[data-testid="time-removed"]').should('contain', '-');
-        cy.get('@row-2').find('[data-testid="helped-by"]').should('contain', '-');
-        cy.get('@row-2').find('[data-testid="removed-by"]').should('contain', '-');
-        cy.get('@row-2').find('[data-testid="removal-method"]').should('contain', '-');
+        checkRows(wisozaRows);
 
         cy.get('#times-helped-cell').should('contain', '0 times helped.');
-
-        // Disable and delete all queue
-        disableQueue();
     });
+
     it('Enabling push and sound notifications as Instructor', () => {
         // Ensure notifications are allowed
         expect(isPermissionAllowed('notifications')).to.be.true;
 
-        cy.login('instructor');
-        enableQueue();
-        cy.visit(['sample', 'office_hours_queue']);
+        switchUser('instructor');
 
         // Assert that switches exist and assign aliases
         cy.get('[data-testid="notification-switch-container"]').first().as('switch-container');
@@ -277,7 +304,13 @@ describe('test office hours queue', () => {
         cy.get('@sound-switch').click();
         cy.get('@push-enabled').should('equal', false);
         cy.get('@audio-enabled').should('equal', false);
+    });
 
-        disableQueue();
+    after(() => {
+        cy.login();
+        // Delete all created queues
+        deleteQueue(queueName);
+        deleteQueue(queueName_random);
+        deleteQueue(queueName_blank);
     });
 });
