@@ -158,7 +158,7 @@ describe('Docker UI Test', () => {
             + 'already exists in capability cpp');
     });
 
-    it('Should add new image', () => {
+    it('Should add new image and remove it', () => {
         cy.reload();
         // Add a new image
         cy.get('#capability-form')
@@ -173,5 +173,15 @@ describe('Docker UI Test', () => {
         cy.get('.alert-success')
             .should('have.text', 'submitty/python:2.7 found on DockerHub'
             + ' and queued to be added!');
+
+        cy.reload();
+        // Remove the image
+        cy.get('[data-image-id="submitty/python:2.7"]')
+            .should('be.visible', { timeout: 10000 })
+            .click();
+        // Confirm dialog return true
+        cy.on('window:confirm', () => true);
+        cy.get('[data-image-id="submitty/python:2.7"]')
+            .should('not.exist');
     });
 });
