@@ -1,5 +1,3 @@
-import { getCurrentSemester } from '../../support/utils.js';
-
 function validate_navigation_page_sections(sections) {
     const section_keys = Object.keys(sections);
     cy.get('.course-section-heading').should('have.length', section_keys.length);
@@ -9,15 +7,175 @@ function validate_navigation_page_sections(sections) {
     });
 }
 
-describe('navigation page', () => {
-    beforeEach(() => {
-        cy.visit('/');
-        cy.viewport(1920, 1200);
+const white = 'rgb(255, 255, 255)';
+const blue = 'rgb(0, 99, 152)';
+const red = 'rgb(220, 53, 69)';
+const green = 'rgb(95, 183, 96)';
+const buttonOrder = ['team-btn', 'submit-btn', 'grade-btn', 'quick-link-btn'];
+
+function checkButtons(gradeableID, buttonText, buttonColor) {
+    // assert that the course does not exist
+    if (buttonText === null && buttonColor === null) {
+        cy.get(`[data-testid="${gradeableID}"]`).should('not.exist');
+    }
+    else {
+        for (let i = 0; i < buttonOrder.length; i++) {
+            if (buttonText[i]) {
+                cy.get(`[data-testid="${gradeableID}"]`).find(`[data-testid="${buttonOrder[i]}"]`).should('contain', buttonText[i]);
+                cy.get(`[data-testid="${gradeableID}"]`).find(`[data-testid="${buttonOrder[i]}"]`).should('have.css', 'background-color', buttonColor[i]);
+            }
+            else {
+                // assert that the button does not exist
+                cy.get(`[data-testid="${gradeableID}"]`).find(`[data-testid="${buttonOrder[i]}"]`).should('not.exist');
+            }
+        }
+    }
+}
+
+describe('tests navigation buttons for each level of access', () => {
+    // Future (No TAs) Homework, Future(TAs) Homework, Open Homework, Open Team Homework, Closed Homework, Closed Team Homework
+    // grading homework, Grading Lab
+    // Grade Released Homework, Grade Released Test,
+    it('should display instructor buttons correctly', () => {
+        cy.login('instructor');
+        cy.visit(['sample']);
+
+        checkButtons('future_no_tas_homework',
+            [null, 'ALPHA SUBMIT', 'PREVIEW GRADING', 'OPEN TO TAS NOW'],
+            [null, white, white, blue],
+        );
+        checkButtons('future_tas_homework',
+            [null, 'BETA SUBMIT', 'PREVIEW GRADING', 'OPEN NOW'],
+            [null, white, white, blue],
+        );
+        checkButtons('open_homework',
+            [null, 'SUBMIT', 'PREVIEW GRADING', 'CLOSE SUBMISSIONS NOW'],
+            [null, blue, white, white],
+        );
+        checkButtons('open_team_homework',
+            ['CREATE TEAM', 'MUST BE ON A TEAM TO SUBMIT', 'PREVIEW GRADING', 'CLOSE SUBMISSIONS NOW'],
+            [blue, blue, white, white],
+        );
+        checkButtons('closed_homework',
+            [null, 'LATE SUBMIT', 'PREVIEW GRADING', 'OPEN TO GRADING NOW'],
+            [null, red, white, blue],
+        );
+        checkButtons('closed_team_homework',
+            ['CREATE TEAM', 'MUST BE ON A TEAM TO SUBMIT', 'PREVIEW GRADING', 'OPEN TO GRADING NOW'],
+            [red, red, white, blue],
+        );
+        checkButtons('grading_homework',
+            [null, 'OVERDUE SUBMISSION', 'GRADE', 'RELEASE GRADES NOW'],
+            [null, red, blue, blue],
+        );
+        checkButtons('grading_lab',
+            [null, null, 'GRADE', 'RELEASE GRADES NOW'],
+            [null, null, white, blue],
+        );
+        checkButtons('grades_released_homework_autota',
+            [null, 'VIEW GRADE', 'REGRADE', null],
+            [null, green, white, null],
+        );
+        checkButtons('grades_released_homework',
+            [null, 'OVERDUE SUBMISSION', 'REGRADE', null],
+            [null, red, white, null],
+        );
+        checkButtons('grades_released_lab',
+            [null, null, 'REGRADE', null],
+            [null, null, white, null],
+        );
     });
 
+    it('should display TA buttons correctly', () => {
+        cy.login('ta');
+        cy.visit(['sample']);
+
+        checkButtons('future_no_tas_homework', null, null);
+        checkButtons('future_tas_homework',
+            [null, 'BETA SUBMIT', 'PREVIEW GRADING', null],
+            [null, white, white, null],
+        );
+        checkButtons('open_homework',
+            [null, 'SUBMIT', 'PREVIEW GRADING', null],
+            [null, blue, white, null],
+        );
+        checkButtons('open_team_homework',
+            ['CREATE TEAM', 'MUST BE ON A TEAM TO SUBMIT', 'PREVIEW GRADING', null],
+            [blue, blue, white, null],
+        );
+        checkButtons('closed_homework',
+            [null, 'LATE SUBMIT', 'PREVIEW GRADING', null],
+            [null, red, white, null],
+        );
+        checkButtons('closed_team_homework',
+            ['CREATE TEAM', 'MUST BE ON A TEAM TO SUBMIT', 'PREVIEW GRADING', null],
+            [red, red, white, null],
+        );
+        checkButtons('grading_homework',
+            [null, 'OVERDUE SUBMISSION', 'GRADE', null],
+            [null, red, blue, null],
+        );
+        checkButtons('grading_lab',
+            [null, null, 'GRADE', null],
+            [null, null, white, null],
+        );
+        checkButtons('grades_released_homework_autota',
+            [null, 'OVERDUE SUBMISSION', 'REGRADE', null],
+            [null, red, white, null],
+        );
+        checkButtons('grades_released_homework',
+            [null, 'OVERDUE SUBMISSION', 'REGRADE', null],
+            [null, red, white, null],
+        );
+        checkButtons('grades_released_lab',
+            [null, null, 'REGRADE', null],
+            [null, null, white, null],
+        );
+    });
+
+    it('should display student buttons correctly', () => {
+        cy.login('student');
+        cy.visit(['sample']);
+
+        checkButtons('future_no_tas_homework', null, null);
+        checkButtons('future_tas_homework', null, null);
+        checkButtons('open_homework',
+            [null, 'SUBMIT', null, null],
+            [null, blue, null, null],
+        );
+        checkButtons('open_team_homework',
+            ['MANAGE TEAM', 'SUBMIT', null, null],
+            [blue, blue, null, null],
+        );
+        checkButtons('closed_homework',
+            [null, 'LATE SUBMIT', null, null],
+            [null, red, white, null],
+        );
+        checkButtons('closed_team_homework',
+            ['VIEW TEAM', 'LATE SUBMIT', null, null],
+            [white, red, null, null],
+        );
+        checkButtons('grading_homework',
+            [null, 'OVERDUE SUBMISSION', null, null],
+            [null, red, null, null],
+        );
+        checkButtons('grading_lab', null, null);
+        checkButtons('grades_released_homework_autota',
+            [null, 'OVERDUE SUBMISSION', null, null],
+            [null, red, null, null],
+        );
+        checkButtons('grades_released_homework',
+            [null, 'OVERDUE SUBMISSION', null, null],
+            [null, red, null, null],
+        );
+        checkButtons('grades_released_lab', null, null);
+    });
+});
+
+describe('navigation page', () => {
     it('should show instructor content for instructor', () => {
         cy.login('instructor');
-        cy.visit(`/courses/${getCurrentSemester()}/sample`);
+        cy.visit(['sample']);
 
         const sections = {
             future: 5,
@@ -113,7 +271,7 @@ describe('navigation page', () => {
 
     it('should show full access grader content for ta', () => {
         cy.login('ta');
-        cy.visit(`/courses/${getCurrentSemester()}/sample`);
+        cy.visit(['sample']);
 
         const sections = {
             beta: 3,
@@ -130,7 +288,7 @@ describe('navigation page', () => {
 
     it('should show student content for student', () => {
         cy.login('student');
-        cy.visit(`/courses/${getCurrentSemester()}/sample`);
+        cy.visit(['sample']);
 
         const sections = {
             open: 8,
