@@ -1650,7 +1650,7 @@ function enableKeyToClick() {
 }
 
 function peerFeedbackUpload(grader_id, user_id, g_id, feedback) {
-    $('#save_status').html('Saving Feedback...');
+    $('#save_status').text('Saving Feedback...').css('color', 'var(--text-black)');
     const url = buildCourseUrl(['gradeable', g_id, 'feedback', 'set']);
     const formData = new FormData();
     formData.append('csrf_token', csrfToken);
@@ -1666,15 +1666,15 @@ function peerFeedbackUpload(grader_id, user_id, g_id, feedback) {
         contentType: false,
         success: function (data) {
             if (data.status === 'success') {
-                $('#save_status').html('All Changes Saved');
+                $('#save_status').text('All Changes Saved').css('color', 'var(--text-black)');
             }
             else {
-                $('#save_status').html('Error Saving Changes');
+                $('#save_status').text('Error Saving Changes').css('color', 'red');
             }
         },
         error: function () {
             window.alert('Something went wrong. Please try again.');
-            $('#save_status').html('<span style="color: red">Some Changes Failed!</span>');
+            $('#save_status').text('Some Changes Failed!').css('color', 'red');
         },
     });
 }
