@@ -225,8 +225,6 @@ function changeStudentArrowTooltips(data) {
             data = 'default';
         }
     }
-    console.log(inquiry_status);
-    console.log(data);
     let component_id = NO_COMPONENT_ID;
     switch (data) {
         case 'ungraded':
@@ -302,7 +300,6 @@ const orig_toggleComponent = window.toggleComponent;
 window.toggleComponent = function (component_id, saveChanges) {
     const ret = orig_toggleComponent(component_id, saveChanges);
     return ret.then(() => {
-        console.log(localStorage.getItem('general-setting-arrow-function'));
         changeStudentArrowTooltips(localStorage.getItem('general-setting-arrow-function') || 'default');
     });
 };
@@ -647,10 +644,16 @@ function updateCookies() {
 // -----------------------------------------------------------------------------
 // Student navigation
 
-function waitForAllAjaxToComplete() {
-    if ($.active > 0) {
-        setTimeout(waitForAllAjaxToComplete, 100);
-    }
+function waitForAllAjaxToComplete(callback) {
+    const checkAjax = () => {
+        if ($.active > 0) {
+            setTimeout(checkAjax, 100);
+        }
+        else {
+            callback();
+        }
+    };
+    checkAjax();
 }
 
 function gotoMainPage() {
@@ -658,8 +661,9 @@ function gotoMainPage() {
 
     if (getGradeableId() !== '') {
         closeAllComponents(true).then(() => {
-            waitForAllAjaxToComplete();
-            window.location = window_location;
+            waitForAllAjaxToComplete(() => {
+                window.location = window_location;
+            });
         }).catch(() => {
             if (confirm('Could not save open component, go to main page anyway?')) {
                 window.location = window_location;
@@ -717,8 +721,9 @@ function gotoPrevStudent() {
 
     if (getGradeableId() !== '') {
         closeAllComponents(true).then(() => {
-            waitForAllAjaxToComplete();
-            window.location = window_location;
+            waitForAllAjaxToComplete(() => {
+                window.location = window_location;
+            });
         }).catch(() => {
             if (confirm('Could not save open component, change student anyway?')) {
                 window.location = window_location;
@@ -776,8 +781,9 @@ function gotoNextStudent() {
 
     if (getGradeableId() !== '') {
         closeAllComponents(true).then(() => {
-            waitForAllAjaxToComplete();
-            window.location = window_location;
+            waitForAllAjaxToComplete(() => {
+                window.location = window_location;
+            });
         }).catch(() => {
             if (confirm('Could not save open component, change student anyway?')) {
                 window.location = window_location;
